### 2025-05-05 #702 @HzaCode
[Pre-submission Inquiry]: Athlytics
Submitting Author Name: Zhiang HE
Submitting Author Github Handle: HzaCode
Other Package Authors Github handles:
Repository:  https://github.com/HzaCode/Athlytics
Submission type: Pre-submission

---

```
Package: Athlytics
Title: Advanced Sports Performance Analysis for 'Strava' Data
Version: 0.1.0
Author: Zhiang HE [aut, cre]
Maintainer: Zhiang HE <ang@hezhiang.com>
    person(given = "Zhiang", family = "HE", email = "ang@hezhiang.com", role = c("aut", "cre"))
Description: Provides a suite of tools for advanced sports performance analysis and modeling, designed to work with activity data retrieved from 'Strava'. It focuses on applying established sports science models and statistical methods to gain deeper insights into training load, performance prediction, recovery status, and identifying key performance factors, extending basic data analysis capabilities.
    httr,
    rlang (>= 0.4.0),
    rStrava,
    viridis,
    zoo,
    mockery
    devtools,
    pkgdown,
    mockery
NeedsCompilation: no
```

- Please indicate which category or categories from our package fit policies or statistical package categories this package falls under. (Please check one or more appropriate boxes below):

`Athlytics` performs `data retrieval` from the Strava API and subsequent `data munging` to prepare activity data. Its core function is applying established **sports science models** for `Time Series Analyses` of performance trends and physiological responses. It facilitates `Exploratory Data Analysis` by computing and visualizing these specialized metrics, providing an integrated `workflow automation` from raw data to advanced insights, going beyond basic descriptive statistics.

- If submitting a statistical package, have you already incorporated documentation of standards into your code via the **srr** package?
No, not yet.

The target audience includes athletes, coaches, sports scientists, and researchers analyzing Strava data. Scientific applications involve studying training effectiveness, modeling performance, and investigating physiological responses to exercise reproducibly.

-   Are there other R packages that accomplish the same thing? If so, how does yours differ or meet our criteria for best-in-category?
    Yes, related packages exist, but `Athlytics` offers distinct value:
    -   `rStrava` mainly gets data from the Strava API. `Athlytics` uses this data (or similar methods) but focuses on *analyzing* it with sports science models (like training load).
    -   `trackeR` is great for analyzing activity *files* (GPX, TCX) and calculating general stats. `Athlytics` focuses on *API data* from Strava and applies specific sports science models not found in `trackeR`.

    `Athlytics` fills a specific niche by integrating Strava API data directly with these advanced, validated sports science modeling techniques within R, offering analytical capabilities not directly provided by pure API wrappers or general file analyzers, thus making sophisticated performance analysis more accessible and reproducible for the R community.

-   (If applicable) Does your package comply with our guidance around _Ethics, Data Privacy and Human Subjects Research_?
The package processes personal activity data retrieved via user authentication from Strava, relying on user authorization and Strava's platform for privacy controls.

No. I'm submitting this as a pre-submission inquiry mainly to ensure `Athlytics` aligns well with rOpenSci's goals and scope before investing further effort into preparing for the full review process. Perhaps the automated package checks (`check package`) might not be necessary.Any initial thoughts on its fit would be very helpful at this point.Thank you!
