### 2025-05-07 1be053e
Update NEWS.md content
