### 2025-05-07 a338e33
Apply package updates (R, man, tests, config files)
