+++ tests/testthat/test-calculate_decoupling.R
@@ -2,72 +2,106 @@
 
 library(testthat)
 library(Athlytics)
-library(rStrava)
+library(mockery)
+# library(rStrava) # Not calling real rStrava functions in this test file
 
-# --- Helper: Dummy stoken ---
-dummy_stoken <- structure(list(endpoint = NULL, app = NULL, credentials = list(dummy = "token")), class = "Token2.0")
+# Load data: sample data from package & mock API returns from helper
+data(Athlytics_sample_data)
+source(test_path("helpe-mockdata.R"), local = TRUE) # Provides mock_activity_list_list, mock_activity_streams
 
-# --- Test Cases for calculate_decoupling ---
+# Mock Strava token - needed for function signature but API calls will be mocked
+mock_stoken <- structure(
+  list(token = list(access_token = "fake_token")), class = "Token2.0"
+)
 
-test_that("calculate_decoupling returns a data frame with expected columns", {
-  skip_on_cran()
-  # This function relies heavily on stream data, making it hard to test without mocks
-  # Basic structure check might be possible if the function handles cases
-  # where no suitable activities are found gracefully (e.g., returns empty df)
+# --- Test calculate_decoupling (API Path Logic with Mocking) ---
+
+test_that("calculate_decoupling (API path) processes mocked data correctly", {
   
-  # Attempt a run, expecting it might return an empty df or fail gracefully
-  # if dummy token prevents actual data fetch needed for streams.
-  # A better approach requires mocking rStrava::get_streams
-  result_df <- tryCatch({
-      calculate_decoupling(
-        stoken = dummy_stoken,
-        activity_type = "Run",
-        decouple_metric = "Pace_HR",
-        max_activities = 2, # Minimal activities for test structure
-        min_duration_mins = 1 # Low duration for test
-      )
-    }, 
-    error = function(e) {
-      # If it errors due to API call failure (expected without mock/real token), return NULL for test check
-      message("API call likely failed as expected without mock/real token: ", e$message)
-      return(NULL) 
+  # Check if necessary mock objects exist
+  expect_true(exists("mock_activity_list_list"), "mock_activity_list_list not found in helper.")
+  expect_true(exists("mock_activity_streams"), "mock_activity_streams not found in helper.")
+  
+  # Mock the rStrava functions called internally by calculate_decoupling
+  local_mocked_bindings(
+    .package = "rStrava",
+    # Mock getting the list of activities
+    get_activity_list = function(stoken, ...) {
+      # message("MOCK get_activity_list called") # Optional debug message
+      return(mock_activity_list_list) 
+    },
+    # Mock getting the streams for ANY activity ID requested
+    # We return the *same* mock stream data regardless of the ID requested,
+    # simplifying the test while still checking the processing logic.
+    get_activity_streams = function(act_data, acts, stoken, types, resolution, ...) {
+      # message(sprintf("MOCK get_activity_streams called for index %d", acts)) # Optional debug message
+      # Input validation based on expected types from calculate_decoupling
+      expect_true(is.list(act_data))
+      expect_true(is.numeric(acts))
+      expect_s3_class(stoken, "Token2.0")
+      expect_true(is.character(types))
+      # Return the standard mock stream data
+      return(mock_activity_streams)
     }
   )
 
-  # Only proceed with checks if the function didn't error out completely
-  if (!is.null(result_df)) {
-      expect_s3_class(result_df, "data.frame")
-      # Check core columns, names might vary slightly
-      expect_true(all(c("activity_id", "start_date", "activity_type", "decoupling_metric", "decoupling_percent") %in% names(result_df)))
[…]
