### 2025-05-07 c6223be
Add package data and documentation files
