### 2025-05-07 380d886
Stage all current changes for R CMD check readiness
