### 2025-05-07 0d24520
Delete tests/manual_tests directory
