+++ tests/manual_tests/run_all_tests.R
@@ -1,188 +1,0 @@
-# --- Test Script for Athlytics Package Functions ---
-
-# --- 0. Load Package ---
-message("Loading Athlytics package using library(Athlytics)...")
-tryCatch({
-  library(Athlytics)
-}, error = function(e) {
-  stop("Error loading package with library(Athlytics). Is the package installed? Error: ", e$message)
-})
-
-# --- 1. Load Required Libraries ---
-message("Loading required libraries (rStrava, httr, dplyr)...")
-tryCatch({
-  library(rStrava)
-  library(httr) # Needed for Token object check
-  library(dplyr) # Often used internally or helpful for checks
-}, error = function(e) {
-  stop("Error loading required libraries (rStrava, httr, dplyr). Are they installed? Error: ", e$message)
-})
-
-cat("Libraries loaded.\n")
-
-# --- 2. Strava Authentication ---
-# Attempts authentication. Stops script on failure.
-message("Performing Strava Authentication...")
-app_name <- 'MyAthlyticsApp'
-# Use environment variables or secure methods in real CI/automated scenarios
-client_id <- Sys.getenv("STRAVA_CLIENT_ID", "118977") # Default if env var not set
-client_secret <- Sys.getenv("STRAVA_CLIENT_SECRET", "1bc6d6c3c26c114e7479ba561adadbf7faaa8057") # Default if env var not set
-
-stoken <- tryCatch({
-  cat("Attempting Strava authentication using rStrava::strava_oauth()...\n")
-  cat("Using Client ID: ", client_id, "\n") # Log the ID being used
-  cat("Check browser if prompted for authorization.\n")
-  rStrava::strava_oauth(app_name = app_name,
-                        app_client_id = client_id, # Use correct param name
-                        app_secret = client_secret, # Use correct param name
-                        app_scope = "activity:read_all",
-                        cache = TRUE) # Cache the token for future use
-}, error = function(e) {
-  stop("Strava authentication failed: ", e$message) # Stop script on error
-})
-
-if (!is.null(stoken) && inherits(stoken, "Token2.0")) {
-    message("Authentication successful or valid cached token loaded.")
-} else {
-    # This case might not be reachable if stop() always works on error
-    stop("Authentication process did not return a valid token object.")
-}
-
-# --- 3. Define Test Parameters ---
-message("Defining test parameters...")
-# Use a reasonable number for testing, avoid hitting rate limits excessively
-test_max_activities <- 5 # Reducing from 10 to 5 for faster testing
-user_ftp <- 250
-user_max_hr <- 190
-user_resting_hr <- 60
-
-message(sprintf("Using max_activities = %d for relevant tests.", test_max_activities))
-message(sprintf("Using FTP = %d, Max HR = %d, Resting HR = %d where needed.", user_ftp, user_max_hr, user_resting_hr))
-
-
-# --- 4. Run Tests ---
-# Each block wrapped in tryCatch to allow continuation if one test fails
-# -- Test: calculate_exposure (TSS) and plot_exposure --
-message("\n--- Testing: calculate_exposure (TSS) & plot_exposure ---")
-tryCatch({
-  exposure_res_tss <- calculate_exposure(
-    stoken = stoken, activity_type = "Run",
-    load_metric = "tss", user_ftp = user_ftp,
-    acute_period = 7, chronic_period = 28
-  )
-  message("calculate_exposure (TSS) finished. Result preview:")
-  print(head(exposure_res_tss))
[…]
