+++ tests/manual_tests/manual_test_script.R
@@ -1,474 +1,0 @@
-# Strava API v3 Base URL: https://www.strava.com/api/v3
-
-library(Athlytics) # Load package using library()
-library(rStrava)   # Load rStrava for authentication and API functions
-library(httr)      # Load httr for potential OOB auth / lower-level control
-
-rm(list = ls(pattern = "^calculate_")) # Remove potential conflicts
-app_name    = "api_testing"    # Match the name set in Strava settings
-app_client_id   = "118977"     # User-provided Client ID
-app_secret = "c13f03c13dc139e11b8a67c186ee055d98b04ce0" # 请妥善保管你的 Secret!
-
-# --- Strava Authentication ---
-# Strava API Application settings page (to get Client ID/Secret):
-# https://www.strava.com/settings/api
-# IMPORTANT: Replace placeholders with your actual Strava API credentials if different!
-
-message("\n--- Attempting Strava Authentication (will use cache if available) ---")
-
-# 尝试使用缓存加载或进行认证
-stoken <- tryCatch({
-  rStrava::strava_oauth(
-    app_name = app_name,
-    app_client_id = app_client_id,
-    app_secret = app_secret,
-    app_scope = "activity:read_all",
-    cache = TRUE # <-- 设置为 TRUE 以启用缓存加载/保存
-  )
-}, error = function(e) {
-  message("ERROR during Strava authentication: ", e$message)
-  message("Please ensure your Strava API credentials (Client ID, Secret) are correct and you complete the browser authentication if prompted.")
-  NULL # 返回 NULL 表示失败
-})
-
-# Check if token was successfully obtained
-if (is.null(stoken) || !inherits(stoken, "Token2.0")) {
-  stop("Failed to obtain a valid Strava token. Please check credentials and authentication process.")
-} else {
-  message("Successfully obtained Strava token.")
-}
-
-# --- Define a global limit for max activities in tests ---
-# Adjust this value to control how many recent activities are fetched for testing
-# Increase this to improve chances of getting 10 successful decoupling results
-test_max_activities <- 50 # Increase attempt count, e.g., to 50
-message(sprintf("\n--- Using a maximum of %d activities for relevant tests --- \n", test_max_activities))
-
-# ============================================================
-#  Test: calculate_exposure (using duration_mins)
-# ============================================================
-# --- Calculation ---
-exposure_data_duration <- tryCatch({
-  calculate_exposure(stoken,
-    activity_type = c("Run", "Ride"), # Test with multiple types
-    load_metric = "duration_mins",
-    acute_period = 7,
-    chronic_period = 42,
-    end_date = Sys.Date() # Use today as end date
-  )
-}, error = function(e) {
-  message("ERROR during calculate_exposure (duration): ", e$message)
-  NULL
-})
-
-# --- Results & Plotting ---
-if (exists("exposure_data_duration") && !is.null(exposure_data_duration) && nrow(exposure_data_duration) > 0) {
-  print("Exposure data (duration):")
-  print(head(exposure_data_duration)) # Print head for brevity
-  message("Plotting Exposure (Duration)...")
-  tryCatch({
-    plot_exposure(exposure_df = exposure_data_duration) # Pass the calculated df
-  }, error = function(e) {
-    message("Error plotting exposure (duration): ", e$message)
-  })
-} else {
-  message("Skipping Exposure (Duration) plot: Calculation failed or data not available.")
-}
-
-# ============================================================
[…]
