### 2025-05-08 7b51c51
Update R CMD check to find package dynamically and add package tarball
