### 2025-05-08 1735dde
Fix: Use direct R CMD check command for tarball
