### 2025-05-08 42fac5e
Debug: Temporarily skip examples in CI for windows-latest due to platform-specific hanging issue
