+++ tests/testthat/test-acwr.R
@@ -15,7 +15,7 @@ data(Athlytics_sample_data)
 # and is essential for testing the raw calculation logic of calculate_acwr.
 # Consider refactoring helper-mockdata.R to only contain truly raw API mocks if needed,
 # or integrate such mocks into a different structure if Athlytics_sample_data contains only processed data.
-source(test_path("helpe-mockdata.R"), local = TRUE) # Assuming this file name is correct and it defines mock_activity_list_list
+source(test_path("helper-mockdata.R"), local = TRUE) # Assuming this file name is correct and it defines mock_activity_list_list
 
 # Mock Strava token - use structure() to correctly set class attributes
 mock_stoken <- structure(
@@ -80,14 +80,11 @@ test_that("calculate_acwr output structure and basic numeric checks with mocked 
   # ACWR can be positive or NA/Inf if CTL is zero, already handled by is.finite
 })
 
-test_that("calculate_acwr handles invalid inputs", {
+test_that("calculate_acwr input parameter validations", {
+  # Existing invalid stoken test
   expect_error(calculate_acwr(stoken = "invalid"), ".*Token2.0 object.*")
 
-  local_mocked_bindings(
-    .package = "rStrava",
-    get_activity_list = function(...) mock_activity_list_list 
-  )
-
+  # Existing acute_period vs chronic_period test (ensure mock_stoken is used here)
   expect_error(
     calculate_acwr(
       stoken = mock_stoken,
@@ -97,6 +94,231 @@ test_that("calculate_acwr handles invalid inputs", {
       end_date = test_end_date
     ),
     message = "`acute_period` must be less than `chronic_period`."
+  )
+
+  # New tests
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, acute_period = -1),
+    message = "`acute_period` must be a positive integer."
+  )
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, chronic_period = 0),
+    message = "`chronic_period` must be a positive integer."
+  )
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, load_metric = "tss", user_ftp = NULL),
+    message = "`user_ftp` is required when `load_metric` is 'tss'."
+  )
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, load_metric = "hrss", user_max_hr = NULL, user_resting_hr = 100),
+    message = "`user_max_hr` and `user_resting_hr` are required when `load_metric` is 'hrss'."
+  )
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, load_metric = "hrss", user_max_hr = 200, user_resting_hr = NULL),
+    message = "`user_max_hr` and `user_resting_hr` are required when `load_metric` is 'hrss'."
+  )
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, load_metric = "invalid_metric"),
+    message = "Invalid `load_metric`."
+  )
+  expect_error(
+    calculate_acwr(stoken = mock_stoken, start_date = "2023-10-01", end_date = "2023-09-01"),
+    message = "start_date must be before end_date."
+  )
+})
+
+test_that("calculate_acwr works with load_metric = distance_km", {
+  expect_true(exists("mock_activity_list_list"))
+  local_mocked_bindings(
+    .package = "rStrava",
+    get_activity_list = function(...) mock_activity_list_list
+  )
+
+  result_df <- calculate_acwr(
+    stoken = mock_stoken,
+    load_metric = "distance_km", # Test distance metric
+    activity_type = NULL, 
+    acute_period = 7,
+    chronic_period = 14, # Use shorter periods for test
[…]
