+++ tests/testthat/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(Athlytics)
+
+test_check("Athlytics", filter = "^(test-acwr|test-ef)") 
