+++ tests/testthat/test-calculate_exposure.R
@@ -6,7 +6,10 @@ library(mockery)
 
 # Load data: sample data from package & mock API returns from helper
 data(Athlytics_sample_data)
-source(test_path("helpe-mockdata.R"), local = TRUE) # Provides mock_activity_list_list
+source(test_path("helper-mockdata.R"), local = TRUE)
+
+# Load mock activity list for testing calculate_exposure's internal processing
+source(test_path("helper-mockdata.R"), local = TRUE) # Assuming this file defines mock_activity_list_list
 
 # Mock Strava token
 mock_stoken <- structure(list(token = list(access_token = "fake_token")), class = "Token2.0")
@@ -89,4 +92,194 @@ test_that("calculate_exposure calculates correctly with mock activity data", {
   # ACWR can be NA/Inf if ctl is 0
   expect_true(all(is.finite(result_df$acwr) | is.na(result_df$acwr)))
   
+})
+
+test_that("calculate_exposure works with load_metric = distance_km", {
+  expect_true(exists("mock_activity_list_list"))
+  local_mocked_bindings(
+    .package = "rStrava",
+    get_activity_list = function(...) mock_activity_list_list
+  )
+  test_end_date <- lubridate::as_date("2023-10-01") 
+  test_chronic_period <- 14
+  test_acute_period <- 7
+
+  result_df <- calculate_exposure(
+    stoken = mock_stoken,
+    activity_type = NULL, 
+    load_metric = "distance_km", # Test distance metric
+    acute_period = test_acute_period,
+    chronic_period = test_chronic_period,
+    end_date = test_end_date
+  )
+  
+  # Basic structure and type checks (similar to duration test)
+  expect_s3_class(result_df, "data.frame")
+  expect_named(result_df, c("date", "daily_load", "atl", "ctl", "acwr"), ignore.order = TRUE)
+  expect_s3_class(result_df$date, "Date")
+  expect_true(is.numeric(result_df$daily_load))
+  expect_true(is.numeric(result_df$atl))
+  expect_true(is.numeric(result_df$ctl))
+  expect_true(is.numeric(result_df$acwr))
+  # Check if daily_load values seem plausible based on mock distance data (mock has distances like 5050m, 20100m etc.)
+  # Example: Expect some loads around 5, 10, 20 etc.
+  expect_true(any(result_df$daily_load > 1))
+  expect_true(all(result_df$daily_load >= 0, na.rm = TRUE))
+})
+
+test_that("calculate_exposure works with load_metric = elevation_gain_m", {
+  expect_true(exists("mock_activity_list_list"))
+  local_mocked_bindings(
+    .package = "rStrava",
+    get_activity_list = function(...) mock_activity_list_list
+  )
+  test_end_date <- lubridate::as_date("2023-10-01")
+  test_chronic_period <- 14
+  test_acute_period <- 7
+
+  result_df <- calculate_exposure(
+    stoken = mock_stoken,
+    activity_type = NULL,
+    load_metric = "elevation_gain_m", # Test elevation metric
+    acute_period = test_acute_period,
+    chronic_period = test_chronic_period,
+    end_date = test_end_date
+  )
+
+  # Basic structure and type checks
+  expect_s3_class(result_df, "data.frame")
+  expect_named(result_df, c("date", "daily_load", "atl", "ctl", "acwr"), ignore.order = TRUE)
+  expect_s3_class(result_df$date, "Date")
+  expect_true(is.numeric(result_df$daily_load))
+  expect_true(is.numeric(result_df$atl))
+  expect_true(is.numeric(result_df$ctl))
+  expect_true(is.numeric(result_df$acwr))
[…]
