+++ NEWS.md
@@ -1,3 +1,10 @@
+# Athlytics 1.1.2
+
+*   **Testing**: Focused on increasing test coverage towards the goal of 85% across the package, with significant improvements made to `R/calculate_acwr.R`. Integrated Codecov for ongoing coverage monitoring.
+*   **Bug Fixes**: Addressed a bug in the test suite (`tests/testthat/test-acwr.R`) by correcting expectations for error handling in `calculate_acwr()` when invalid parameters are provided.
+
+---
+
 # Athlytics 0.1.1
 
 ## Core Improvement: Enhanced Reliability & Testing with Simulated Data
