+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: Athlytics
 Title: Advanced Sports Performance Analysis for 'Strava' Data
-Version: 0.1.1
+Version: 1.1.2
 Author: Ang [aut, cre]
 Maintainer: Ang <ang@hezhiang.com>
 Authors@R: 
