+++ man/plot_acwr.Rd
@@ -68,7 +68,7 @@ if (!is.null(athlytics_sample_acwr)) {
   print(p)
 }
 
-\donttest{
+\dontrun{
 # Example using real data (requires authentication)
 # Please replace with your actual Strava application details for this to work.
 # stoken_example <- rStrava::strava_oauth(
