+++ R/calculate_ef.R
@@ -29,7 +29,7 @@
 #' data(Athlytics_sample_data)
 #' print(head(athlytics_sample_ef))
 #'
-#' \donttest{
+#' \dontrun{
 #' # Example using real data (requires authentication)
 #' # To authenticate (replace with your details):
 #' # stoken <- rStrava::strava_oauth(app_name = "YOUR_APP",
