+++ R/plot_exposure.R
@@ -44,7 +44,7 @@
 #' p <- plot_exposure(exposure_df = athlytics_sample_exposure, activity_type = "Run")
 #' print(p)
 #'
-#' \donttest{
+#' \dontrun{
 #' # Example using real data (requires authentication)
 #' # stoken <- rStrava::strava_oauth("YOUR_APP_NAME",
 #' #                                "YOUR_APP_CLIENT_ID",
