+++ man/calculate_pbs.Rd
@@ -46,7 +46,7 @@ Provides data for `plot_pbs`. Processes activities chronologically.
 data(Athlytics_sample_data)
 print(head(athlytics_sample_pbs))
 
-\donttest{
+\dontrun{
 # Example using real data (requires authentication and YOUR credentials)
 # NOTE: The following rStrava::strava_oauth call is a placeholder.
 # You MUST replace "YOUR_APP_NAME", "YOUR_CLIENT_ID", "YOUR_CLIENT_SECRET"
