### 2025-05-10 2412b19
Delete run_coverage.R
