### 2025-05-10 02badc9
Update NEWS.md based on CRAN feedback and regenerate HTML
