+++ paper/paper.bib
@@ -1,0 +1,284 @@
+
+@Manual{R-base,
+  title        = {{R: A Language and Environment for Statistical Computing}},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address      = {Vienna, Austria},
+  year         = {2025},
+  url          = {https://www.R-project.org/},
+}
+
+@Manual{R-rStrava,
+  title        = {{rStrava: Access the 'Strava' API}},
+  author       = {Marcus W. Beck and Pedro Villarroel and Daniel Padfield and Lorenzo Gaborini and Niklas von Maltzahn},
+  year         = {2024},
+  note         = {R package version 1.3.2},
+  url          = {https://CRAN.R-project.org/package=rStrava},
+}
+
+@Book{ggplot2-book,
+  author    = {Hadley Wickham},
+  title     = {{ggplot2: Elegant Graphics for Data Analysis}},
+  publisher = {Springer International Publishing},
+  year      = {2016},
+  edition   = {2nd},
+  isbn      = {978-3-319-24277-4},
+  doi       = {10.1007/978-3-319-24277-4},
+  url       = {https://link.springer.com/book/10.1007/978-3-319-24277-4},
+}
+
+@Manual{R-dplyr,
+  title = {{dplyr: A Grammar of Data Manipulation}},
+  author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller and Davis Vaughan},
+  year = {2023},
+  note = {R package version 1.1.4},
+  url = {https://dplyr.tidyverse.org},
+}
+
+@article{lubridate-jss,
+  title   = {{Dates and Times Made Easy with {lubridate}}},
+  author  = {Garrett Grolemund and Hadley Wickham},
+  journal = {Journal of Statistical Software},
+  year    = {2011},
+  volume  = {40},
+  number  = {3},
+  pages   = {1--25},
+  url     = {https://www.jstatsoft.org/v40/i03/},
+  doi     = {10.18637/jss.v040.i03},
+}
+
+@article{zoo-jss,
+  title   = {{zoo: S3 Infrastructure for Regular and Irregular Time Series}},
+  author  = {Achim Zeileis and Gabor Grothendieck},
+  journal = {Journal of Statistical Software},
+  year    = {2005},
+  volume  = {14},
+  number  = {6},
+  pages   = {1--27},
+  doi     = {10.18637/jss.v014.i06},
+  url     = {https://www.jstatsoft.org/v014/i06/},
+}
+
+@Manual{R-rlang,
+  title = {{rlang: Functions for Base Types and Core R and 'Tidyverse' Features}},
+  author = {Lionel Henry and Hadley Wickham},
+  year = {2025},
+  note = {R package version 1.1.6},
+  url = {https://rlang.r-lib.org},
+}
+
+@Manual{R-purrr,
+  title = {{purrr: Functional Programming Tools}},
+  author = {Hadley Wickham and Lionel Henry},
+  year = {2025},
+  note = {R package version 1.0.4},
+  url = {https://purrr.tidyverse.org},
+}
+
+@Manual{R-remotes,
+  title = {{remotes: R Package Installation from Remote Repositories, Including 'GitHub'}},
+  author = {Gábor Csárdi and Jim Hester and Hadley Wickham and Winston Chang and Martin Morgan and Dan Tenenbaum},
+  year = {2024},
+  note = {R package version 2.5.0},
+  url = {https://remotes.r-lib.org},
+}
+
+@article{trackeR-jss,
+  title   = {{trackeR: Infrastructure for Running and Cycling Data from GPS-Enabled Tracking Devices in R}},
[…]
