### 2025-05-18 c52b643
Delete LICENSE.md
