### 2025-05-18 af0e7ae
Update calculate_acwr.R
