+++ R/calculate_acwr.R
@@ -163,7 +163,7 @@ calculate_acwr <- function(stoken,
         return(NULL)
     }
     if (!is.null(activity_type) && !act_type %in% activity_type) {
-        message("    -> Filtered out by type.")
+        # message("    -> Filtered out by type.")
         return(NULL)
     }
 
@@ -175,12 +175,12 @@ calculate_acwr <- function(stoken,
     elevation_gain <- safe_as_numeric(.x$total_elevation_gain)
     # Use device_watts if available (direct power), otherwise fallback to average_watts (estimated power)
     np_proxy <- safe_as_numeric(.x$device_watts %||% .x$average_watts %||% 0) 
-    message(sprintf("    Duration: %.0f sec", duration_sec))
+    # message(sprintf("    Duration: %.0f sec", duration_sec))
 
     # --- Added Debugging and Refined Logic --- 
-    message(sprintf("    Inputs check: load_metric='%s', duration_sec=%.1f, distance_m=%.1f, avg_hr=%.1f, np_proxy=%.1f, user_ftp=%s, user_max_hr=%s, user_resting_hr=%s", 
-                    load_metric, duration_sec, distance_m, avg_hr, np_proxy, 
-                    deparse(user_ftp), deparse(user_max_hr), deparse(user_resting_hr)))
+    # message(sprintf("    Inputs check: load_metric='%s', duration_sec=%.1f, distance_m=%.1f, avg_hr=%.1f, np_proxy=%.1f, user_ftp=%s, user_max_hr=%s, user_resting_hr=%s", 
+    #                 load_metric, duration_sec, distance_m, avg_hr, np_proxy, 
+    #                 deparse(user_ftp), deparse(user_max_hr), deparse(user_resting_hr)))
     
     if (duration_sec > 0) {
       # Initialize load_value outside case_when to handle default case cleanly
@@ -202,7 +202,7 @@ calculate_acwr <- function(stoken,
             avg_hr_rel <- (avg_hr - user_resting_hr) / hr_reserve
             load_value <- (duration_sec / 60) * avg_hr_rel # Simplified TRIMP
           } else {
-              message("    Skipping HRSS calculation: Missing/invalid HR parameters or avg_hr out of range.")
+              # message("    Skipping HRSS calculation: Missing/invalid HR parameters or avg_hr out of range.")
           }
       } else if (load_metric == "tss") {
            # Check required FTP parameter before calculating
@@ -210,25 +210,25 @@ calculate_acwr <- function(stoken,
              intensity_factor <- np_proxy / user_ftp
              load_value <- (duration_sec * np_proxy * intensity_factor) / (user_ftp * 3600) * 100
            } else {
-               message("    Skipping TSS calculation: Missing/invalid FTP or power data (np_proxy).")
+               # message("    Skipping TSS calculation: Missing/invalid FTP or power data (np_proxy).")
            }
       }
       
-      message(sprintf("    Calculated load_value: %.2f", load_value))
+      # message(sprintf("    Calculated load_value: %.2f", load_value))
     } else {
-        message("    Duration <= 0, load is 0.")
+        # message("    Duration <= 0, load is 0.")
         load_value <- 0 # Ensure load_value is defined even if duration is 0
     }
 
     if (!is.na(load_value) && load_value > 0) {
-      message("    -> Activity PASSED filters, returning load data.")
+      # message("    -> Activity PASSED filters, returning load data.")
       data.frame(
         date = activity_date,
         load = load_value,
         stringsAsFactors = FALSE
       )
     } else {
[…]
