+++ paper/paper.bib
@@ -1,0 +1,223 @@
+@Manual{R-base,
+title        = {{R: A Language and Environment for Statistical Computing}},
+author       = {{R Core Team}},
+organization = {R Foundation for Statistical Computing},
+address      = {Vienna, Austria},
+year         = {2025},
+url          = {https://www.R-project.org/},
+}
+
+@article{Gabbett2016,
+  author = {T. J. Gabbett},
+  title = {{The training—injury prevention paradox: should athletes be training smarter and harder?}},
+  journal = {British Journal of Sports Medicine},
+  year = {2016},
+  volume = {50},
+  number = {5},
+  pages = {273--280},
+  doi = {10.1136/bjsports-2015-095788},
+  url = {https://doi.org/10.1136/bjsports-2015-095788}
+}
+
+@article{Hulin2016,
+  author = {B. T. Hulin and T. J. Gabbett and D. W. Lawson and P. Caputi and J. A. Sampson},
+  title = {{The acute: chronic workload ratio predicts injury: high chronic workload may decrease injury risk in elite rugby league players}},
+  journal = {British Journal of Sports Medicine},
+  year = {2016},
+  volume = {50},
+  number = {4},
+  pages = {231--236},
+  doi = {10.1136/bjsports-2015-094817},
+  url = {https://doi.org/10.1136/bjsports-2015-094817}
+}
+
+@book{allen2019training,
+  title     = {{Training and Racing with a Power Meter}},
+  author    = {Hunter Allen and Andrew R. Coggan and Stephen McGregor},
+  isbn      = {9781937715939},
+  publisher = {VeloPress},
+  year      = {2019},
+  edition   = {3rd},
+  url       = {https://www.amazon.com/Training-Racing-Power-Meter-Hunter/dp/1937715930},
+  doi       = {No DOI assigned for this specific book edition}
+}
+
+@article{Maunder2021,
+  author  = {Ed Maunder and Stephen Seiler and Mathew J. Mildenhall and Andrew E. Kilding and Daniel J. Plews},
+  title   = {{The Importance of 'Durability' in the Physiological Profiling of Endurance Athletes}},
+  journal = {Sports Medicine},
+  year    = {2021},
+  volume  = {51},
+  number  = {8},
+  pages   = {1619--1628},
+  doi     = {10.1007/s40279-021-01459-0},
+  url     = {https://doi.org/10.1007/s40279-021-01459-0}
+}
+
+@article{tidyverse-joss,
+  author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy {D'Agostino McGowan} and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana P. Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus O. Wilke and Kara Woo and Hiroaki Yutani},
+  title = {{Welcome to the {tidyverse}}},
+  journal = {Journal of Open Source Software},
+  year = {2019},
+  volume = {4},
+  number = {43},
+  pages = {1686},
+  doi = {10.21105/joss.01686},
+  url = {https://joss.theoj.org/papers/10.21105/joss.01686}
+}
+
+@article{Bourdon2017,
[…]
