+++ paper/paper.bib
@@ -192,7 +192,7 @@ url          = {https://www.R-project.org/},
   title   = {{Athlytics: A Computational Framework for Longitudinal Analysis of Exercise Physiology Metrics from Wearable Sensor Data}},
   author  = {Zhiang He},
   year    = {2025},
-  note    = {R package version 1.0.0. Available from CRAN.},
+  note    = {R package version 0.1.2. Available from CRAN.},
   url     = {https://CRAN.R-project.org/package=Athlytics},
   doi     = {10.32614/CRAN.package.Athlytics}
 }
