+++ paper/paper.bib
@@ -189,7 +189,7 @@ url          = {https://www.R-project.org/},
 }
 
 @Manual{R-Athlytics,
-  title   = {{Athlytics: A Computational Framework for Longitudinal Analysis of Exercise Physiology Metrics from Wearable Sensor Data}},
+  title   = {{Athlytics: Advanced Sports Performance Analysis for 'Strava' Data}},
   author  = {Zhiang He},
   year    = {2025},
   note    = {R package version 0.1.2. Available from CRAN.},
