+++ paper/paper.md
@@ -59,7 +59,7 @@ Understanding dynamic physiological responses to exercise is central to exercise
 
 ## Comparison with Similar Software
 
-Within the R ecosystem for sports science analytics, `Athlytics` carves out a distinct niche through its integrated analytical workflow specifically designed for Strava API data and a curated set of physiological metrics. While the `rStrava` package [@R-rStrava] provides the essential channel for researchers to access Strava data, `Athlytics` builds a critical **analytical layer** upon it, directly transforming raw activity data into physiological insights suitable for longitudinal research. Compared to more broadly functional R packages like `trackeR` [@trackeR-jss]—which excels in handling diverse local sports tracking files (e.g., .tcx, .gpx) and offering general-purpose sports science analytics—`Athlytics` focuses on the direct utilization of Strava API data and provides **ready-to-use modules** for core physiological indicators of interest to researchers (such as ACWR, EF, and decoupling). This design obviates the need for researchers working with Strava data to undertake cumbersome data integration and workflow concatenation, as might be required when using multiple single-function R packages (e.g., a dedicated package for ACWR calculation, like `ACWR` [@R-ACWR] if a specific citation exists, or custom scripts). `Athlytics` offers a **coherent framework from data acquisition to the calculation and visualization of multi-dimensional physiological metrics**. Adhering to `tidyverse` design principles, it not only lowers the technical barrier for researchers conducting such specific analyses but, more importantly, significantly enhances the **reproducibility and methodological transparency** of exercise physiology research based on wearable device data by providing an open-source, standardized computational environment, directly addressing the core values of scientific software.
+Within the R ecosystem for sports science analytics, `Athlytics` carves out a distinct niche through its integrated analytical workflow specifically designed for Strava API data and a curated set of physiological metrics. While the `rStrava` package [@R-rStrava] provides the essential channel for researchers to access Strava data, `Athlytics` builds a critical **analytical layer** upon it, directly transforming raw activity data into physiological insights suitable for longitudinal research. Compared to more broadly functional R packages like `trackeR` [@trackeR-jss]—which excels in handling diverse local sports tracking files (e.g., .tcx, .gpx) and offering general-purpose sports science analytics—`Athlytics` focuses on the direct utilization of Strava API data and provides **ready-to-use modules** for core physiological indicators of interest to researchers (such as ACWR, EF, and decoupling). This design obviates the need for researchers working with Strava data to undertake cumbersome data integration and workflow concatenation, as might be required when using multiple single-function R packages (e.g., a dedicated package for ACWR calculation like `ACWR` [@R-ACWR], or custom scripts). `Athlytics` offers a **coherent framework from data acquisition to the calculation and visualization of multi-dimensional physiological metrics**. Adhering to `tidyverse` design principles, it not only lowers the technical barrier for researchers conducting such specific analyses but, more importantly, significantly enhances the **reproducibility and methodological transparency** of exercise physiology research based on wearable device data by providing an open-source, standardized computational environment, directly addressing the core values of scientific software.
 
 
 
