### 2025-05-20 #702 @mpadge
Thanks for you inquiry @HzaCode and apologies for taking a while to get back to you. I'll send a request for automated package checks now, and proceed to more specific issues following that. I anticipate that we'll have some conceptual issues regarding scientific application of this package, so maybe you could help in the meantime with a brief answer to the following:

- I can see the immediate utility of the functionality provided by {Athlytics} in helping me (as a passionate cyclist 🚴 ) in my personal training regime, but have a harder time conceiving of broader utility for scientific research. Could you please help me, and potential editors and reviewers to come, to understand in more detail how the outputs of {Athlytics} help to advance the _scientific_ processes of sports analytics?

Part of a potential problem that I foresee in that question is that the Strava API only offers access to my personal data, and only to others with their explicit approval. The package therefore seems inherently restricted to only being able to analyse _my_ data in all their fullness, while being unable to position the outputs within any broader analytic context of data from others, or even any kind of general or average data. Do you have any ideas on how I might contextualize output from your package based on my personal data within more general scientific understandings of the metrics you provide?
