### 2025-05-22 #702 @HzaCode
Hi @mpadge, thanks for your feedback.

In terms of broader data use: group-level analysis is possible when multiple users authorize access, which is how team-based or collaborative athlete studies can be done.As for benchmarking features and broader context, those are definitely on my list and something I plan to explore when I have more time.
