+++ paper/paper.bib
@@ -39,7 +39,6 @@ url          = {https://www.R-project.org/},
   year      = {2019},
   edition   = {3rd},
   url       = {https://www.amazon.com/Training-Racing-Power-Meter-Hunter/dp/1937715930},
-  doi       = {No DOI assigned for this specific book edition}
 }
 
 @article{Maunder2021,
