### 2025-06-24 #702 @mpadge
Thank you for the kind words @HzaCode. I'll close this issue for now, but please feel very welcome to re-open at any stage in future to continue discussions. Good luck in the meantime with the package.
