### 2025-07-28 9c9d484
Update paper.md
