+++ paper/paper.md
@@ -36,7 +36,7 @@ The proliferation of wearable sensors and platforms like Strava has generated va
 
 ## Statement of Need
 
-Understanding dynamic physiological responses to exercise is central to exercise science and personalized health [@Bourdon2017]. While wearable sensors generate rich data streams [@Hicks2019analyzing], established open-source workflows in R for systematically quantifying key physiological indicators (e.g., ACWR, EF, decoupling) from popular platform APIs like Strava are lacking. This often forces researchers to expend considerable effort on custom programming for data retrieval, metric calculation, and longitudinal visualization, limiting analytical scale, efficiency, and reproducibility [@Sanders2017]. `Athlytics` addresses this critical bottleneck by providing a dedicated R framework. It seamlessly integrates data acquisition (via the `rStrava` package [@R-rStrava]) with the calculation and visualization of these exercise physiology metrics. This empowers broader research applications by enabling researchers to efficiently test hypotheses regarding the dynamic interplay between training stimuli, physiological efficiency, and stress responses using ubiquitous data sources, while carefully considering the necessary approximations for certain composite load metrics.
+Understanding dynamic physiological responses to exercise is central to exercise science and personalized health [@Bourdon2017]. While wearable sensors generate rich data streams [@Hicks2019analyzing], and platforms like Strava serve as a de facto data repository for a vast population of athletes, established open-source workflows in R for systematically quantifying key physiological indicators (e.g., ACWR, EF, decoupling) from its API are lacking. This methodological gap often forces researchers to expend considerable effort on custom programming for data retrieval, metric calculation, and longitudinal visualization, limiting analytical scale, efficiency, and reproducibility [@Sanders2017]. Athlytics addresses this critical bottleneck by providing a dedicated R framework. By requiring only a one-time authentication token per athlete, it grants researchers programmatic and continuous access to longitudinal data, thereby overcoming a primary hurdle in data acquisition. The package seamlessly integrates data acquisition (via the rStrava package [@R-rStrava]) with the calculation and visualization of these exercise physiology metrics. This empowers broader research applications by enabling researchers to efficiently test hypotheses regarding the dynamic interplay between training stimuli, physiological efficiency, and stress responses using ubiquitous data sources, while carefully considering the necessary approximations for certain composite load metrics.
 
 ## Key Functionalities
 
@@ -61,71 +61,43 @@ Understanding dynamic physiological responses to exercise is central to exercise
 
 ## Comparison with Similar Software
 
[…]
