+++ README.md
@@ -11,305 +11,189 @@ Athlytics <img src="https://github.com/HzaCode/Athlytics/blob/main/image.png?raw
 [![Website](https://img.shields.io/badge/website-Athlytics-blue)](https://hezhiang.com/Athlytics/)
 [![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg)](CONTRIBUTING.md)
 [![bioRxiv](https://img.shields.io/badge/bioRxiv-Preprint-F88A0B)](https://www.biorxiv.org/content/10.1101/2025.05.01.651597v2)
-### Overview
-
-This is the development repository for **Athlytics**, an R package designed to simplify the analysis of athletic performance and training load data sourced **directly from the Strava API**. The package provides functions for fetching your Strava data, calculating key metrics, and generating insightful visualizations for training monitoring.
-
-### Recent Releases
-
-For details on recent changes and version history, please see the [Changelog](https://hezhiang.com/Athlytics/news/index.html) on the package website.
-<div style="padding: 15px; border: 2px solid #0073B7; background-color: #f0f8ff; margin-bottom: 20px; border-radius: 5px;">
-  <h3 style="margin-top: 0; color: #005a8c;">🚀 Install Released Version from CRAN 📦</h3>
-  <p>
-    The <strong>stable, released version</strong> of <code>Athlytics</code> (currently <strong>0.1.2</strong>) is available on <a href="https://CRAN.R-project.org/package=Athlytics" target="_blank" rel="noopener noreferrer">CRAN</a>.
-  </p>
-  <p>
-    ✅ <strong>This is the highly recommended way to install <code>Athlytics</code> for most users:</strong>
-  </p>
-  <p>👉 Simply copy and paste the following line into your R console:</p>
-  <pre style="background-color: #e9ecef; padding: 10px; border: 1px solid #ced4da; overflow-x: auto; border-radius: 3px;"><code class="language-r" style="color: #212529;">install.packages("Athlytics")</code></pre>
-  <p style="font-size: 0.9em; color: #555;">
- 
-  </p>
-</div>
 
 
+**Athlytics** is an R package that provides a computational framework for the longitudinal analysis of exercise physiology metrics. It is specifically designed to streamline the acquisition, processing, and visualization of data directly from the Strava API. The package addresses a critical need in sports science research by providing a standardized, reproducible workflow to quantify key indicators of training load, performance, and physiological adaptation, such as the Acute:Chronic Workload Ratio (ACWR), Efficiency Factor (EF), and cardiovascular decoupling. By interfacing directly with Strava—a nearly ubiquitous platform for athletes—`Athlytics` enables researchers to efficiently analyze rich, real-world data to investigate the dynamic interplay between training stimuli and human physiology.
 
-Alternatively, the development version from this repository can be installed as follows:
+### Installation
 
+The stable version of `Athlytics` is available on CRAN and can be installed with:
+```r
+install.packages("Athlytics")
+```
+
+Alternatively, the latest development version can be installed from GitHub:
 ```r
 # install.packages('remotes')
 remotes::install_github('HzaCode/Athlytics')
 ```
 
+### Authentication Protocol
 
-### Using Athlytics Functions
+`Athlytics` leverages the `rStrava` package for handling the OAuth2.0 authentication process with the Strava API. This requires a one-time setup of a Strava API application to obtain a Client ID and Secret. The resulting token object is then passed to `Athlytics` functions to authorize data access.
 
-```r
-library(Athlytics)
-```
-
-### Authentication with Strava (using `rStrava`)
-
[…]
