### 2025-07-28 607b03c
Update paper.md
