### 2025-07-28 a6d319e
Update paper.md
