### 2025-07-28 92b3cfb
Update paper.md
