+++ R/zzz.R
@@ -6,10 +6,10 @@
   pkg_version <- pkg_desc$Version
   
   startup_msg <- paste0(
-    "\nLoading Athlytics version ", pkg_version, ".\n",
+    "\nLoading athlytics version ", pkg_version, ".\n",
     "Analyze your Strava data with ease!\n",
     "Remember to authenticate using rStrava::strava_oauth() first.\n",
-    "For documentation, see: https://hzacode.github.io/Athlytics/"
+    "For documentation, see: https://hzacode.github.io/athlytics/"
   )
 
 

[…]
