+++ README.md
@@ -1,199 +1,139 @@
-athlytics <img src="https://github.com/HzaCode/athlytics/blob/main/image.png?raw=true" align="right" width="200"/>
-=======================================================
-
+# Athlytics <img src="https://github.com/HzaCode/athlytics/blob/main/image.png?raw=true" align="right" width="180"/>
 
 [![Project Status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-![CRAN status](https://www.r-pkg.org/badges/version/Athlytics)
 [![R-CMD-check](https://github.com/HzaCode/athlytics/actions/workflows/R-CMD-check.yml/badge.svg)](https://github.com/HzaCode/athlytics/actions/workflows/R-CMD-check.yml)
 [![codecov](https://codecov.io/gh/HzaCode/athlytics/graph/badge.svg)](https://app.codecov.io/gh/HzaCode/athlytics)
-![](https://img.shields.io/badge/lifecycle-experimental-orange.svg)
-[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT) 
+![CRAN status](https://www.r-pkg.org/badges/version/Athlytics)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![Website](https://img.shields.io/badge/website-athlytics-blue)](https://hezhiang.com/athlytics/)
 [![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg)](CONTRIBUTING.md)
 [![bioRxiv](https://img.shields.io/badge/bioRxiv-Preprint-F88A0B)](https://www.biorxiv.org/content/10.1101/2025.05.01.651597v3)
 
+---
 
-**athlytics** is an R package that provides a computational framework for the longitudinal analysis of exercise physiology metrics. It is specifically designed to streamline the acquisition, processing, and visualization of data directly from the Strava API. The package addresses a critical need in sports science research by providing a standardized, reproducible workflow to quantify key indicators of training load, performance, and physiological adaptation, such as the Acute:Chronic Workload Ratio (ACWR), Efficiency Factor (EF), and cardiovascular decoupling. By interfacing directly with Strava—a nearly ubiquitous platform for athletes—`athlytics` enables researchers to efficiently analyze rich, real-world data to investigate the dynamic interplay between training stimuli and human physiology.
+`Athlytics` is an R framework for the longitudinal analysis of exercise physiology using Strava data. It is designed to remove the primary bottleneck for sports scientists: the programmatic acquisition and consistent processing required to turn raw, high-frequency ecological data into quantitative insights.
 
-### Installation
+---
+### ✨ Core Features
 
-The stable version of `athlytics` is available on CRAN and can be installed with:
+*   **📈 Longitudinal Quantification**
+    Automated calculation of configurable metrics (ACWR, EF). Enables longitudinal tracking of physiological adaptation.
+
+*   **💔 High-Resolution Decoupling**
+    Robust cardiovascular drift analysis from raw stream data. Methodologically superior to metrics from simple averages.
+
+*   **🧹 Tidy & Modeling-Ready Outputs**
+    Clean `tibbles` from every function. Eliminates data cleaning. Primed for `ggplot2` visualization and statistical modeling.
+
[…]
