### 2025-08-08 3fd4328
Fix CRAN license stub: make LICENSE a stub and move full MIT text to LICENSE.md
