### 2025-08-08 bc9d946
CI: build tarball, derive package name from DESCRIPTION, and upload .Rcheck artifact
