### 2025-08-08 ef20537
CI: robust tarball discovery, clean old artifacts, dynamic package load before coverage
