+++ .github/workflows/R-CMD-check.yml
@@ -38,6 +38,11 @@ jobs:
           extra-packages: any::rcmdcheck, any::covr, any::devtools
           needs: check
 
+      - name: Clean old tarballs
+        run: |
+          rm -f *.tar.gz || true
+        shell: bash
+
       - name: Build source package
         run: |
           R CMD build .
@@ -54,7 +59,13 @@ jobs:
       - name: Find R package tarball
         id: find_tarball
         run: |
-          TARBALL_NAME=$(ls *.tar.gz 2>/dev/null | head -n 1)
+          PKG="${{ steps.pkg_info.outputs.PKG }}"
+          # Prefer newest tarball for the current package name
+          TARBALL_NAME=$(ls -t ${PKG}_*.tar.gz 2>/dev/null | head -n 1)
+          # Fallback: any newest tarball
+          if [ -z "$TARBALL_NAME" ]; then
+            TARBALL_NAME=$(ls -t *.tar.gz 2>/dev/null | head -n 1)
+          fi
           if [ -z "$TARBALL_NAME" ]; then
             echo "Error: No .tar.gz file found in the root directory for R CMD check."
             exit 1
@@ -75,14 +86,18 @@ jobs:
         if: success() # Only run if R CMD check succeeds
         shell: Rscript {0}
         run: |
-          message("Installing package from tarball for coverage testing: ${{ steps.find_tarball.outputs.TARBALL_NAME }}")
+          pkg <- "${{ steps.pkg_info.outputs.PKG }}"
+          tarball <- "${{ steps.find_tarball.outputs.TARBALL_NAME }}"
+          message(sprintf("Installing package from tarball for coverage testing: %s", tarball))
           if (!requireNamespace("devtools", quietly = TRUE)) install.packages("devtools")
-          devtools::install_local("${{ steps.find_tarball.outputs.TARBALL_NAME }}", force = TRUE, quiet = TRUE)
+          devtools::install_local(tarball, force = TRUE, quiet = TRUE)
 
-          message("Ensuring covr and testthat are available for run_coverage.R")
           if (!requireNamespace("covr", quietly = TRUE)) install.packages("covr")
           if (!requireNamespace("testthat", quietly = TRUE)) install.packages("testthat")
-          
+
+          message(sprintf("Loading package: %s", pkg))
+          library(pkg, character.only = TRUE)
+
           message("Executing run_coverage.R to generate and upload coverage...")
           source("run_coverage.R")
 
