+++ paper/paper.md
@@ -31,43 +31,38 @@ doi: null
 
 ## Summary
 
-The proliferation of wearable sensors and consumer platforms like Strava has generated vast longitudinal data streams, creating unprecedented opportunities to investigate dynamic human physiological responses to exercise. However, translating this raw data into scientifically rigorous insights for exercise physiology research faces considerable methodological and computational challenges. `Athlytics` is an R \[@R-base] package designed as a computational framework to streamline the acquisition, analysis, and visualization of key exercise physiology metrics directly from Strava API data. It enables researchers and practitioners to systematically quantify indicators such as Acute\:Chronic Workload Ratio (ACWR) \[@Gabbett2016; @Hulin2016], Efficiency Factor (EF) \[@allen2019training], cardiovascular decoupling \[@Maunder2021], and track personal bests (PBs). By providing standardized function interfaces adhering to `tidyverse` principles \[@tidyverse-joss], `Athlytics` significantly lowers the technical barrier for conducting complex longitudinal analyses, thereby enhancing the feasibility, efficiency, and reproducibility of quantitative research utilizing widely available physiological sensor data.
+The proliferation of wearable sensors and consumer platforms like Strava has generated vast longitudinal data streams, creating unprecedented opportunities to investigate dynamic human physiological responses to exercise. However, translating this raw data into scientifically rigorous insights for exercise physiology research faces considerable methodological and computational challenges. *Athlytics* is an R \[@R-base] package designed as a computational framework to streamline the acquisition, analysis, and visualization of key exercise physiology metrics directly from Strava API data. It enables researchers and practitioners to systematically quantify indicators such as Acute:Chronic Workload Ratio (ACWR) \[@Gabbett2016; @Hulin2016], Efficiency Factor (EF) \[@allen2019training], cardiovascular decoupling \[@Maunder2021], and track personal bests (PBs). By providing standardized function interfaces adhering to `tidyverse` principles \[@tidyverse-joss], *Athlytics* significantly lowers the technical barrier for conducting complex longitudinal analyses, thereby enhancing the feasibility, efficiency, and reproducibility of quantitative research utilizing widely available physiological sensor data.
 
 ## Statement of Need
 
 *Athlytics* targets sports scientists, physiologists, and applied researchers studying longitudinal adaptations to endurance training using consumer wearables and Strava as a data hub. Despite the ubiquity of Strava data access via `rStrava` \[@R-rStrava], there is no standardized, open-source workflow in R that transforms raw API outputs into core physiological indicators (e.g., ACWR, EF, decoupling) with tidy, reproducible interfaces. This methodological gap forces many researchers to spend substantial effort on bespoke scripts for data retrieval, metric calculation, and visualization, limiting analytical scale, efficiency, and methodological transparency \[@Sanders2017].
 
-`Athlytics` addresses this gap by providing an integrated framework from data acquisition to the calculation and visualization of multiple physiological metrics, using paired `calculate_*` and `plot_*` functions built on the `tidyverse` \[@tidyverse-joss]. The package is designed to lower the technical barrier for such analyses while maintaining scientific rigor. It ships with sample datasets, enabling all core functionalities to be demonstrated without requiring Strava API authentication—removing a frequent logistical hurdle when reproducing workflows. By explicitly framing metrics like ACWR as descriptive monitoring tools rather than injury risk predictors, `Athlytics` supports responsible interpretation and application in research.
[…]
