+++ README.md
@@ -1,33 +1,63 @@
-# Athlytics <img src="https://github.com/HzaCode/athlytics/blob/main/image.png?raw=true" align="right" width="180"/>
-![CRAN Version](https://img.shields.io/cran/v/Athlytics?style=for-the-badge&color=276DC3&logo=r&logoColor=white)
+<div align="center">
 
-[![Project Status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![R-CMD-check](https://github.com/HzaCode/athlytics/actions/workflows/R-CMD-check.yml/badge.svg)](https://github.com/HzaCode/athlytics/actions/workflows/R-CMD-check.yml)
-[![codecov](https://codecov.io/gh/HzaCode/athlytics/graph/badge.svg)](https://app.codecov.io/gh/HzaCode/athlytics)
-[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
-[![Website](https://img.shields.io/badge/website-athlytics-blue)](https://hezhiang.com/athlytics/)
-[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg)](CONTRIBUTING.md)
-[![bioRxiv](https://img.shields.io/badge/bioRxiv-Preprint-F88A0B)](https://www.biorxiv.org/content/10.1101/2025.05.01.651597v3)
+<img src="https://github.com/HzaCode/athlytics/blob/main/image.png?raw=true" width="220"/>
+
+# Athlytics
+
+*An R Framework for Longitudinal Analysis of Exercise Physiology*
+
+</div>
+
+<p align="center">
+    <a href="https://cran.r-project.org/package=Athlytics">
+        <img src="https://img.shields.io/cran/v/Athlytics?style=for-the-badge&color=276DC3&logo=r&logoColor=white" alt="CRAN Version">
+    </a>
+</p>
+
+<p align="center">
+    <a href="https://www.repostatus.org/#active">
+        <img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status" style="vertical-align: middle;">
+    </a>
+    <a href="https://github.com/HzaCode/athlytics/actions/workflows/R-CMD-check.yml">
+        <img src="https://github.com/HzaCode/athlytics/actions/workflows/R-CMD-check.yml/badge.svg" alt="R-CMD-check" style="vertical-align: middle;">
+    </a>
+    <a href="https://app.codecov.io/gh/HzaCode/athlytics">
+        <img src="https://codecov.io/gh/HzaCode/athlytics/graph/badge.svg" alt="Codecov" style="vertical-align: middle;">
+    </a>
+    <a href="https://opensource.org/licenses/MIT">
+        <img src="https://img.shields.io/badge/License-MIT-yellow.svg" alt="License: MIT">
+    </a>
+</p>
+
+<p align="center">
+    <a href="https://hezhiang.com/athlytics/">
+        <img src="https://img.shields.io/badge/website-athlytics-blue" alt="Website">
+    </a>
+    <a href="CONTRIBUTING.md">
+        <img src="https://img.shields.io/badge/PRs-welcome-brightgreen.svg" alt="PRs Welcome">
+    </a>
+    <a href="https://www.biorxiv.org/content/10.1101/2025.05.01.651597v3">
+        <img src="https://img.shields.io/badge/bioRxiv-Preprint-F88A0B" alt="bioRxiv Preprint">
+    </a>
+</p>
 
 ---
 
[…]
