### 2025-09-24 33f4556
Standardize package name to Athlytics
