+++ R/data.R
@@ -1,6 +1,6 @@
 # R/data.R
 
-#' Sample ACWR Data for athlytics
+#' Sample ACWR Data for Athlytics
 #'
 #' A dataset containing pre-calculated Acute:Chronic Workload Ratio (ACWR)
 #' and related metrics, derived from simulated Strava data. Used in examples and tests.
@@ -16,7 +16,7 @@
 #' @source Simulated data generated for package examples.
 "athlytics_sample_acwr"
 
-#' Sample Aerobic Decoupling Data for athlytics
+#' Sample Aerobic Decoupling Data for Athlytics
 #'
 #' A dataset containing pre-calculated aerobic decoupling percentages,
 #' derived from simulated Strava data. Used in examples and tests.
@@ -29,7 +29,7 @@
 #' @source Simulated data generated for package examples.
 "athlytics_sample_decoupling"
 
-#' Sample Efficiency Factor (EF) Data for athlytics
+#' Sample Efficiency Factor (EF) Data for Athlytics
 #'
 #' A dataset containing pre-calculated Efficiency Factor (EF) values,
 #' derived from simulated Strava data. Used in examples and tests.
@@ -43,7 +43,7 @@
 #' @source Simulated data generated for package examples.
 "athlytics_sample_ef"
 
-#' Sample Training Load Exposure Data for athlytics
+#' Sample Training Load Exposure Data for Athlytics
 #'
 #' This dataset contains daily training load, ATL, CTL, and ACWR, derived from
 #' simulated Strava data. Used in examples and tests, particularly for `plot_exposure`.
@@ -59,7 +59,7 @@
 #' @source Simulated data generated for package examples.
 "athlytics_sample_exposure"
 
-#' Sample Personal Bests (PBs) Data for athlytics
+#' Sample Personal Bests (PBs) Data for Athlytics
 #'
 #' A dataset containing pre-calculated Personal Best (PB) times for various distances,
 #' derived from simulated Strava data. Used in examples and tests.
