### 2025-09-25 3e061ce
Fix test files to use correct package name Athlytics
