+++ tests/testthat/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
-library(athlytics)
+library(Athlytics)
 
 # test_check("athlytics", filter = "^(test-acwr|test-ef)") 
