### 2025-09-25 88b211f
Improve CI: Build package from source instead of using local tarball
