### 2025-09-25 5130c94
Fix GitHub Actions: skip vignettes, add pandoc
