### 2025-09-25 4d2dbf1
Fix GitHub Actions workflow with error handling for tests
