### 2025-09-25 fa268f4
Fix segfault: replace complex mocking tests with simple functionality tests
