### 2025-09-25 5b04713
Use CRAN-like workflow: R CMD check first, then basic verification
