### 2025-09-25 1712ace
Fix workflow: CRAN-like check first, then basic tests
