### 2025-09-25 dbff11e
Simplify: only R CMD check needed, remove redundant tests
