### 2025-09-25 061a5ee
chore: Update badges and streamline CI workflows
