+++ README.md
@@ -19,7 +19,9 @@
 <p align="center">
     </a>
     <a href="https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml">
-        <img src="https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml/badge.svg" alt="R-CMD-check" style="vertical-align: middle;">
+        <img src="https://img.shields.io/badge/R--CMD--check-grey?style=flat-square" alt="R-CMD-check" style="vertical-align: middle; margin-right: -4px;">
+        <img src="https://img.shields.io/badge/Windows%20%7C%20macOS%20%7C%20Ubuntu-blue?style=flat-square" alt="Platforms" style="vertical-align: middle; margin-right: -4px;">
+        <img src="https://img.shields.io/github/actions/workflow/status/HzaCode/Athlytics/R-CMD-check.yml?branch=main&label=&style=flat-square" alt="Status" style="vertical-align: middle;">
     </a>
     <a href="https://app.codecov.io/gh/HzaCode/Athlytics">
         <img src="https://codecov.io/gh/HzaCode/Athlytics/graph/badge.svg" alt="Codecov" style="vertical-align: middle;">
@@ -28,10 +30,7 @@
         <img src="https://img.shields.io/badge/License-MIT-yellow.svg" alt="License: MIT">
     </a>
     <a href="https://hezhiang.com/Athlytics/">
-        <img src="https://img.shields.io/badge/website-Athlytics-blue" alt="Website">
-    </a>
-    <a href="CONTRIBUTING.md">
-        <img src="https://img.shields.io/badge/PRs-welcome-brightgreen.svg" alt="PRs Welcome">
+        <img src="https://img.shields.io/badge/docs-passing-brightgreen" alt="Documentation">
     </a>
     <a href="https://www.biorxiv.org/content/10.1101/2025.05.01.651597v3">
         <img src="https://img.shields.io/badge/bioRxiv-Preprint-F88A0B" alt="bioRxiv Preprint">
