+++ .github/workflows/R-CMD-check-minimal.yml
@@ -1,88 +1,0 @@
-on:
-  push:
-    branches: [main, master]
-  pull_request:
-    branches: [main, master]
-
-name: R-CMD-check-minimal
-
-jobs:
-  R-CMD-check:
-    runs-on: ubuntu-latest
-    name: Minimal R Check
-
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-      _R_CHECK_CRAN_INCOMING_: false
-      _R_CHECK_CRAN_INCOMING_REMOTE_: false
-      _R_CHECK_FORCE_SUGGESTS_: false
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: 'release'
-          use-public-rspm: true
-
-      - name: Install system dependencies
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev
-
-      - name: Install basic R dependencies
-        run: |
-          R -e "install.packages(c('devtools', 'testthat', 'mockery'), repos = 'https://cloud.r-project.org')"
-
-      - name: Install package dependencies
-        run: |
-          R -e "
-          # Install required packages from DESCRIPTION
-          install.packages(c('dplyr', 'ggplot2', 'httr', 'jsonlite', 'lubridate', 'purrr', 'rlang', 'rStrava', 'tidyr', 'viridis', 'zoo'))
-          "
-
-      - name: Install package from source (no vignettes)
-        run: |
-          R -e "devtools::install('.', build_vignettes = FALSE, dependencies = FALSE)"
-
-      - name: Run tests
-        run: |
-          R -e "
-          library(testthat)
-          library(Athlytics)
-          library(mockery)
-          
-          # Load test data
-          data(athlytics_sample_data)
-          
-          # Run tests
-          result <- test_dir('tests/testthat', reporter = 'summary')
-          print(result)
-          
-          # Check for failures
-          if (any(as.data.frame(result)[['failed']] > 0)) {
-            stop('Some tests failed')
-          }
-          "
-
-      - name: Basic package check (no vignettes, no examples)
-        run: |
-          R -e "
-          # Build minimal tarball
-          system('R CMD build . --no-build-vignettes --no-manual')
-          
-          # Find the tarball
-          tarball <- list.files(pattern = '*.tar.gz')[1]
-          cat('Found tarball:', tarball, '\n')
-          
-          # Run minimal check
-          system(paste('R CMD check', tarball, '--no-manual --no-build-vignettes --no-examples --no-tests'))
-          "
-
-      - name: Upload check results
-        if: failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: minimal-check-results
-          path: "*.Rcheck"
