### 2025-09-25 d6bf69d
fix(docs): Remove gap in multi-part badge
