### 2025-09-29 b80abd7
feat: improve test coverage to 87.54%
- Add comprehensive test cases for plot_ef, calculate_decoupling, plot_exposure, and strava_helpers
- Fix missing scale_color_viridis_d usage in plot_ef.R
- Add edge case testing for various input scenarios
- Improve error handling test coverage
- All tests pass with 492 test cases
