+++ tests/testthat/test-ef.R
@@ -201,4 +201,512 @@ test_that("plot_ef handles empty data frame input", {
     expect_s3_class(p_empty, "ggplot") # Should still return a ggplot object
     # Check if the plot title indicates no data
     expect_true(grepl("No EF data available", p_empty$labels$title, ignore.case = TRUE))
-}) 
+})
+
+# Test error handling for missing parameters
+test_that("plot_ef handles missing parameters correctly", {
+  # Test missing stoken when ef_df is not provided
+  expect_error(
+    plot_ef(activity_type = "Run", ef_metric = "Pace_HR"),
+    regexp = "Either 'stoken' or 'ef_df' must be provided"
+  )
+})
+
+# Test with different ef_metric options
+test_that("plot_ef works with different ef_metric options", {
+  # Create sample data for Power_HR metric
+  manual_ef_df_power <- data.frame(
+    date = as.Date(c("2023-01-01", "2023-01-08", "2023-01-15")),
+    ef_value = c(2.5, 2.7, 2.6),
+    activity_type = c("Ride", "Ride", "Ride"),
+    stringsAsFactors = FALSE
+  )
+  
+  p_power <- plot_ef(ef_df = manual_ef_df_power, ef_metric = "Power_HR")
+  expect_s3_class(p_power, "ggplot")
+  
+  # Test with Pace_HR metric
+  manual_ef_df_pace <- data.frame(
+    date = as.Date(c("2023-01-01", "2023-01-08", "2023-01-15")),
+    ef_value = c(0.15, 0.16, 0.155),
+    activity_type = c("Run", "Run", "Run"),
+    stringsAsFactors = FALSE
+  )
+  
+  p_pace <- plot_ef(ef_df = manual_ef_df_pace, ef_metric = "Pace_HR")
+  expect_s3_class(p_pace, "ggplot")
+})
+
+# Test add_trend_line parameter
+test_that("plot_ef handles add_trend_line parameter", {
+  manual_ef_df <- data.frame(
+    date = as.Date(c("2023-01-01", "2023-01-08", "2023-01-15", "2023-01-22")),
+    ef_value = c(0.15, 0.16, 0.155, 0.17),
+    activity_type = c("Run", "Run", "Run", "Run"),
+    stringsAsFactors = FALSE
+  )
+  
+  p_with_trend <- plot_ef(ef_df = manual_ef_df, add_trend_line = TRUE)
+  p_without_trend <- plot_ef(ef_df = manual_ef_df, add_trend_line = FALSE)
+  
+  # Check if trend line is present/absent
+  get_smooth_layers <- function(p) sum(sapply(p$layers, function(l) inherits(l$geom, "GeomSmooth")))
+  expect_equal(get_smooth_layers(p_with_trend), 1)
+  expect_equal(get_smooth_layers(p_without_trend), 0)
+})
+
+# Test with different smoothing methods
+test_that("plot_ef works with different smoothing methods", {
+  manual_ef_df <- data.frame(
+    date = as.Date(c("2023-01-01", "2023-01-08", "2023-01-15", "2023-01-22")),
+    ef_value = c(0.15, 0.16, 0.155, 0.17),
+    activity_type = c("Run", "Run", "Run", "Run"),
+    stringsAsFactors = FALSE
+  )
+  
+  p_loess <- plot_ef(ef_df = manual_ef_df, smoothing_method = "loess")
+  p_lm <- plot_ef(ef_df = manual_ef_df, smoothing_method = "lm")
+  
+  expect_s3_class(p_loess, "ggplot")
+  expect_s3_class(p_lm, "ggplot")
+})
+
+# Test with multiple activity types
[…]
