### 2025-10-06 7a753a6
feat: migrate to local data analysis (v1.0.0) - add file parsing support for FIT/TCX/GPX - restore all calculate functions with local data - remove API dependencies
