### 2025-10-08 25feafd
feat: Add ZIP file support for load_local_activities and update to v1.0.0
- Add direct ZIP archive reading capability in load_local_activities()
- Update DESCRIPTION to version 1.0.0
- Enhance function documentation with ZIP file examples
- Refactor tests to use local data instead of API tokens
- Add test coverage for new load_local_activities function
- Update .gitignore to exclude temporary test files
- Remove deprecated API-based test files

Core functionality verified:
- ZIP file extraction and CSV loading: PASS
- ACWR calculation with local data: PASS
- All 42 tests passing with 0 failure
[…]
