+++ tests/testthat/test-acwr.R
@@ -4,356 +4,156 @@ context("ACWR Calculation and Plotting")
 
 library(Athlytics)
 library(testthat)
-library(mockery)
 
 # Load sample data from the package
 data(athlytics_sample_data)
 
-# Load mock activity list for testing calculate_acwr's internal processing
-# This specific mock is for simulating rStrava::get_activity_list output
-# Keep this source line if mock_activity_list_list is not part of athlytics_sample_data
-# and is essential for testing the raw calculation logic of calculate_acwr.
-  # Note: helper-mockdata.R contains raw API mocks
-# or integrate such mocks into a different structure if athlytics_sample_data contains only processed data.
-source(test_path("helper-mockdata.R"), local = TRUE) # Assuming this file name is correct and it defines mock_activity_list_list
-
-# Mock Strava token - use structure() to correctly set class attributes
-mock_stoken <- structure(
-  list(
-    token = list(
-      access_token = "fake_access_token",
-      token_type = "Bearer",
-      expires_at = as.integer(Sys.time() + 3600),
-      refresh_token = "fake_refresh_token"
-    )
-  ),
-  class = c("Token2.0", "Token")
-)
-
-# Define test date range
-test_start_date <- "2023-09-13"
-test_end_date <- "2023-10-01"
-expected_rows <- as.numeric(lubridate::ymd(test_end_date) - lubridate::ymd(test_start_date)) + 1
-
-# --- Test calculate_acwr --- 
-
-test_that("calculate_acwr output structure and basic numeric checks with mocked API", {
-  local_mocked_bindings(
-    .package = "rStrava",
-    get_activity_list = function(...) mock_activity_list_list # Uses list of lists from helpe-mockdata.R
+# Create mock activities data for testing
+create_mock_activities <- function(n = 30) {
+  dates <- seq(Sys.Date() - n, Sys.Date(), by = "day")
+  data.frame(
+    id = seq_len(length(dates)),
+    name = paste("Activity", seq_len(length(dates))),
+    type = sample(c("Run", "Ride"), length(dates), replace = TRUE),
+    sport_type = sample(c("Run", "Ride"), length(dates), replace = TRUE),
+    start_date_local = as.POSIXct(dates),
+    date = as.Date(dates),
+    distance = runif(length(dates), 1000, 15000),  # meters
+    moving_time = as.integer(runif(length(dates), 1200, 5400)),  # seconds
+    elapsed_time = as.integer(runif(length(dates), 1200, 5400)),
+    average_heartrate = runif(length(dates), 120, 170),
+    max_heartrate = runif(length(dates), 160, 190),
+    average_watts = runif(length(dates), 150, 250),
+    max_watts = runif(length(dates), 300, 500),
+    elevation_gain = runif(length(dates), 0, 500),
+    average_speed = runif(length(dates), 2, 5),
+    stringsAsFactors = FALSE
   )
+}
 
-  acwr_duration <- calculate_acwr(
-    stoken = mock_stoken,
+# --- Test calculate_acwr with local data ---
+
+test_that("calculate_acwr works with activities_data parameter", {
+  mock_activities <- create_mock_activities(60)
+  
+  acwr_result <- calculate_acwr(
+    activities_data = mock_activities,
     load_metric = "duration_mins",
-    activity_type = "Run", # Specify activity_type for filtering within calculate_acwr
+    activity_type = "Run",
     acute_period = 7,
-    chronic_period = 14,
-    start_date = test_start_date,
-    end_date = test_end_date
-  )
-
-  # Structure checks
[…]
