+++ tests/testthat/test-ef.R
@@ -1,712 +1,64 @@
 # tests/testthat/test-ef.R
 
-library(testthat)
-library(Athlytics)
-library(mockery)
-# library(rStrava) # Not calling real rStrava functions directly in this test file
+context("Efficiency Factor Calculation")
 
-# Load sample data from the package
+library(Athlytics)
+library(testthat)
+
 data(athlytics_sample_data)
 
-# Load mock activity list for testing calculate_ef's internal processing
-source(test_path("helper-mockdata.R"), local = TRUE) # Assuming this file defines mock_activity_list_list
-
-# Mock Strava token - needed for function signature but API calls will be mocked
-mock_stoken <- structure(
-  list(token = list(access_token = "fake_token")), class = "Token2.0"
-)
-
-# --- Test calculate_ef (API Path Logic with Mocking) ---
-
-test_that("calculate_ef (API path) processes mocked Pace_HR data correctly", {
-  # Check necessary mock objects exist
-  expect_true(exists("mock_activity_list_list"), "mock_activity_list_list not found in helper.")
-  expect_true(exists("mock_activity_streams"), "mock_activity_streams not found in helper.")
-
-  local_mocked_bindings(
-    .package = "rStrava",
-    get_activity_list = function(...) mock_activity_list_list,
-    get_activity_streams = function(...) mock_activity_streams
+create_mock_activities <- function(n = 30) {
+  dates <- seq(Sys.Date() - n, Sys.Date(), by = "day")
+  data.frame(
+    id = seq_len(length(dates)),
+    name = paste("Activity", seq_len(length(dates))),
+    type = sample(c("Run", "Ride"), length(dates), replace = TRUE),
+    date = as.Date(dates),
+    distance = runif(length(dates), 5000, 15000),
+    moving_time = as.integer(runif(length(dates), 1800, 5400)),
+    average_heartrate = runif(length(dates), 130, 170),
+    average_watts = runif(length(dates), 180, 250),
+    average_speed = runif(length(dates), 2.5, 4.5),
+    stringsAsFactors = FALSE
   )
+}
 
-  # Print environment of the function being tested
-  message("CALCULATE_EF_TEST_BLOCK_RUNNING")
-  result_df <- calculate_ef(
-    stoken = mock_stoken,
-    activity_type = "Run", # Filter for runs, as in mock_activity_list_list
-    ef_metric = "Pace_HR",
-    start_date = "2023-09-01", # Covers dates in mock_activity_list_list
-    end_date = "2023-10-01",
-    min_duration_mins = 1 # Low duration to include mock activities
-  )
-
-  expect_s3_class(result_df, "data.frame")
-  expect_named(result_df, c("activity_type", "date", "ef_value"), ignore.order = TRUE)
-  if (nrow(result_df) > 0) { # EF might not be calculable for all mock activities
-    expect_s3_class(result_df$date, "Date")
-    # Check 'ef_value' column type, allowing for NAs
-    if ("ef_value" %in% colnames(result_df)) {
-      expect_true(is.numeric(result_df$ef_value), info = "Column 'ef_value' should be numeric (it can be NA).")
-      if(any(!is.na(result_df$ef_value))) { # If there are any non-NA values, check their type
-         expect_type(result_df$ef_value[!is.na(result_df$ef_value)], "double")
-      }
-    } else {
-      fail("Column 'ef_value' is missing from result_df.")
-    }
-  }
[…]
