### 2025-10-08 f06d3db
feat: Add advanced analysis features and enhanced visualizations
- Add EWMA-based ACWR calculation (calculate_acwr_ewma)
- Add cohort reference comparison tools
- Add color palette system for consistent visualizations
- Add data quality flagging functionality
- Enhance all plot functions with improved styling
- Update sample data with new features
- Add comprehensive test coverage for new features
