+++ R/plot_decoupling.R
@@ -158,27 +158,32 @@ plot_decoupling <- function(stoken,
                             max_activities = 50,
                             add_trend_line = TRUE,
                             smoothing_method = "loess",
-                            decoupling_df = NULL) {
+                            decoupling_df = NULL,
+                            group_var = NULL,
+                            group_colors = NULL) {
 
   # Match arg for decouple_metric to ensure only one is used internally if multiple are provided
   decouple_metric_label <- match.arg(decouple_metric)
+
+  # --- Check if first argument is already decoupling data frame ---
+  if (is.data.frame(stoken) && any(c("decoupling_pct", "decoupling_percent", "decoupling") %in% colnames(stoken))) {
+    decoupling_df <- stoken
+  }
 
   # --- Get Data --- 
   # If decoupling_df is not provided, calculate it
   if (is.null(decoupling_df)) {
       # Check if stoken provided when decoupling_df is not
-      if (missing(stoken)) stop("Either 'stoken' or 'decoupling_df' must be provided.")
+      if (missing(stoken)) stop("Either provide decoupling data frame from calculate_decoupling() as first argument, or provide activities_data.")
       
       message("No pre-calculated decoupling_df provided. Calculating data now... (This may take a while)")
       # Call the calculation function
       decoupling_df <- calculate_decoupling(
-          stoken = stoken,
+          activities_data = stoken,
           activity_type = activity_type, # Can be a vector
-          decouple_metric = decouple_metric_label, # Use the matched, single metric
           start_date = start_date,
           end_date = end_date,
-          min_duration_mins = min_duration_mins,
-          max_activities = max_activities
+          min_duration_mins = min_duration_mins
       )
   }
 
@@ -198,15 +203,14 @@ plot_decoupling <- function(stoken,
   plot_data <- decoupling_df %>%
     dplyr::mutate(date = lubridate::as_date(.data$date))
   
+  # --- Check for group variable ---
+  has_groups <- !is.null(group_var) && group_var %in% colnames(plot_data)
+  
   # --- Plotting ---
   message("Generating plot...")
   y_axis_label <- paste("Decoupling (", gsub("_", "/", decouple_metric_label), ") [%]")
   
   # Determine activity type label for the plot title/subtitle
-  # If activity_type was a parameter to calculate_decoupling, it might not be in decoupling_df
-  # If decoupling_df is provided, it might or might not have an activity_type column.
-  # For simplicity, if `activity_type` arg to `plot_decoupling` is length 1, use it.
-  # Otherwise, use a generic term or extract from data if present.
   plot_activity_type_label <- if (length(activity_type) == 1) {
       activity_type
   } else if ("activity_type" %in% names(plot_data) && length(unique(plot_data$activity_type)) == 1) {
@@ -215,26 +219,74 @@ plot_decoupling <- function(stoken,
       "Selected Activities" # Generic if multiple or not clearly defined
   }
 
-
-  p <- ggplot2::ggplot(plot_data, ggplot2::aes(x = .data$date, y = .data$decoupling)) +
-    ggplot2::geom_point(alpha = 0.8, size = 2.5, color = "dodgerblue") +
-    ggplot2::geom_hline(yintercept = 5, linetype="dashed", color="grey70") +
[…]
