+++ R/cohort_reference.R
@@ -1,0 +1,354 @@
+# R/cohort_reference.R
+
+#' Calculate Cohort Reference Percentiles
+#'
+#' Calculates reference percentiles for a metric across a cohort of athletes,
+#' stratified by specified grouping variables (e.g., sport, sex, age band).
+#'
+#' @param data A data frame containing metric values for multiple athletes.
+#'   Must include columns: `date`, `athlete_id`, and the metric column.
+#' @param metric Name of the metric column to calculate percentiles for
+#'   (e.g., "acwr", "acwr_smooth", "ef", "decoupling"). Default "acwr_smooth".
+#' @param by Character vector of grouping variables. Options: "sport", "sex", 
+#'   "age_band", "athlete_id". Default c("sport").
+#' @param probs Numeric vector of probabilities for percentiles (0-1).
+#'   Default c(0.05, 0.25, 0.50, 0.75, 0.95) for 5th, 25th, 50th, 75th, 95th percentiles.
+#' @param min_athletes Minimum number of athletes required per group to calculate
+#'   valid percentiles. Default 5.
+#' @param date_col Name of the date column. Default "date".
+#'
+#' @return A long-format data frame with columns:
+#'   \describe{
+#'     \item{date}{Date}
+#'     \item{...}{Grouping variables (as specified in `by`)}
+#'     \item{percentile}{Percentile label (e.g., "p05", "p25", "p50", "p75", "p95")}
+#'     \item{value}{Metric value at that percentile}
+#'     \item{n_athletes}{Number of athletes contributing to this percentile}
+#'   }
+#'
+#' @details
+#' This function creates cohort-level reference bands for comparing individual
+#' athlete metrics to their peers. Common use cases:
+#' \itemize{
+#'   \item Compare an athlete's ACWR trend to team averages
+#'   \item Identify outliers (athletes outside P5-P95 range)
+#'   \item Track team-wide trends over time
+#' }
+#'
+#' **Important**: Percentile bands represent **population variability**, not
+#' statistical confidence intervals for individual values.
+#'
+#' @importFrom dplyr group_by summarise filter mutate %>% ungroup
+#' @importFrom tidyr pivot_longer
+#' @importFrom stats quantile
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' # Load activities for multiple athletes
+#' athlete1 <- load_local_activities("athlete1_export.zip") %>%
+#'   mutate(athlete_id = "athlete1")
+#' athlete2 <- load_local_activities("athlete2_export.zip") %>%
+#'   mutate(athlete_id = "athlete2")
+#' athlete3 <- load_local_activities("athlete3_export.zip") %>%
+#'   mutate(athlete_id = "athlete3")
+#'
+#' # Combine data
+#' cohort_data <- bind_rows(athlete1, athlete2, athlete3)
+#'
+#' # Calculate ACWR for each athlete
+#' cohort_acwr <- cohort_data %>%
+#'   group_by(athlete_id) %>%
+#'   do(calculate_acwr_ewma(.))
+#'
+#' # Calculate reference percentiles
+#' reference <- cohort_reference(
+#'   cohort_acwr,
+#'   metric = "acwr_smooth",
+#'   by = c("sport"),
+#'   probs = c(0.05, 0.25, 0.5, 0.75, 0.95)
+#' )
+#'
+#' # Plot individual against cohort
+#' plot_with_reference(
+#'   individual = cohort_acwr %>% filter(athlete_id == "athlete1"),
+#'   reference = reference
+#' )
+#' }
+cohort_reference <- function(data,
+                            metric = "acwr_smooth",
+                            by = c("sport"),
[…]
