### 2025-10-08 9237e6b
chore: Remove documentation files from Git tracking
Remove README.md, NEWS.md, paper/, and vignettes/ from Git tracking.
These files will remain local only as per .gitignore configuration.
