+++ paper/paper.md
@@ -1,114 +1,0 @@
----
-title: 'Athlytics: A Computational Framework for Longitudinal Analysis of Exercise Physiology Metrics from Wearable Sensor Data'
-tags:
-  - R
-  - Strava
-  - exercise physiology
-  - physiological monitoring
-  - computational physiology
-  - bioinformatics
-  - wearable sensors
-  - longitudinal data analysis
-  - training load
-  - ACWR
-  - efficiency factor
-  - decoupling
-  - reproducibility
-  - sports science
-authors:
-  - name: Zhiang He
-    affiliation: 1
-affiliations:
-  - name: Independent researcher
-    index: 1  
-date: 15 May 2025 
-bibliography: paper.bib
-repository: https://github.com/HzaCode/Athlytics 
-version: 1.0.0 
-license: MIT 
-doi: null 
----
-
-## Summary
-
-The proliferation of wearable sensors and consumer platforms like Strava has generated vast longitudinal data streams, creating unprecedented opportunities to investigate dynamic human physiological responses to exercise. However, translating this raw data into scientifically rigorous insights for exercise physiology research faces considerable methodological and computational challenges. *Athlytics* is an R \[@R-base] package designed as a computational framework to streamline the acquisition, analysis, and visualization of key exercise physiology metrics directly from Strava API data. It enables researchers and practitioners to systematically quantify indicators such as Acute:Chronic Workload Ratio (ACWR) \[@Gabbett2016; @Hulin2016], Efficiency Factor (EF) \[@allen2019training], cardiovascular decoupling \[@Maunder2021], and track personal bests (PBs). By providing standardized function interfaces adhering to `tidyverse` principles \[@tidyverse-joss], *Athlytics* significantly lowers the technical barrier for conducting complex longitudinal analyses, thereby enhancing the feasibility, efficiency, and reproducibility of quantitative research utilizing widely available physiological sensor data.
-
-## Statement of Need
-
-*Athlytics* targets sports scientists, physiologists, and applied researchers studying longitudinal adaptations to endurance training using consumer wearables and Strava as a data hub. Despite the ubiquity of Strava data access via `rStrava` \[@R-rStrava], there is no standardized, open-source workflow in R that transforms raw API outputs into core physiological indicators (e.g., ACWR, EF, decoupling) with tidy, reproducible interfaces. This methodological gap forces many researchers to spend substantial effort on bespoke scripts for data retrieval, metric calculation, and visualization, limiting analytical scale, efficiency, and methodological transparency \[@Sanders2017].
-
-*Athlytics* addresses this gap by providing an integrated framework from data acquisition to the calculation and visualization of multiple physiological metrics, using paired `calculate_*` and `plot_*` functions built on the `tidyverse` \[@tidyverse-joss]. The package is designed to lower the technical barrier for such analyses while maintaining scientific rigor.
-
-## Key Functionalities
-
-*Athlytics* offers a modular structure, generally featuring paired `calculate_*` and `plot_*` functions for each analytical task, built upon `dplyr` \[@R-dplyr], `ggplot2` \[@ggplot2], and other `tidyverse` packages.
-
-* **Data Acquisition and Preprocessing:** Leverages `rStrava` \[@R-rStrava] for Strava API authentication and data retrieval. Handles date/time operations using `lubridate` \[@lubridate-jss] and rolling calculations with `zoo` \[@zoo-jss].
-* **ACWR Trend Analysis:**
[…]
