### 2025-10-10 a466dbd
Update documentation and add paper folder to gitignore
