+++ R/calculate_ef.R
@@ -1,22 +1,104 @@
 # R/calculate_ef.R
 
-#' Calculate Efficiency Factor (EF) Data
+#' Calculate Efficiency Factor (EF)
 #'
-#' Calculates Efficiency Factor (Pace/HR or Power/HR) from Strava activities.
+#' Computes Efficiency Factor (EF) for endurance activities, quantifying the
+#' relationship between performance output (pace or power) and heart rate.
+#' EF is a key indicator of aerobic fitness and training adaptation (Allen et al., 2019).
+#'
+#' @description
+#' Efficiency Factor measures how much work you perform per unit of cardiovascular
+#' effort. Higher EF indicates better aerobic fitness - you're able to maintain faster
+#' pace or higher power at the same heart rate. Tracking EF over time helps monitor
+#' aerobic base development and training effectiveness.
+#'
+#' **EF Metrics:**
+#' \itemize{
+#'   \item **Pace_HR** (for running): Speed (m/s) / Average HR
+#'     - Higher values = faster pace at same HR = better fitness
+#'   \item **Power_HR** (for cycling): Average Power (watts) / Average HR
+#'     - Higher values = more power at same HR = better fitness
+#' }
+#'
+#' **What Improves EF?**
+#' \itemize{
+#'   \item Aerobic base building (Zone 2 training)
+#'   \item Improved running/cycling economy
+#'   \item Enhanced cardiovascular efficiency
+#'   \item Increased mitochondrial density
+#' }
 #'
 #' @param activities_data A data frame of activities from `load_local_activities()`.
-#'   Must contain columns: date, type, moving_time, distance, average_heartrate, average_watts.
-#' @param activity_type Character vector or single string specifying activity type(s).
-#' @param ef_metric Character string specifying the EF metric ("Pace_HR" or "Power_HR").
-#' @param start_date Optional start date (YYYY-MM-DD string or Date object). Defaults to one year ago.
-#' @param end_date Optional end date (YYYY-MM-DD string or Date object). Defaults to today.
-#' @param min_duration_mins Numeric, minimum activity duration in minutes. Default 20.
+#'   Must contain columns: `date`, `type`, `moving_time`, `distance`, 
+#'   `average_heartrate`, and `average_watts` (for Power_HR metric).
+#' @param activity_type Character vector or single string specifying activity type(s)
+#'   to analyze. Common values: `"Run"`, `"Ride"`, or `c("Run", "Ride")`.
+#'   Default: `c("Run", "Ride")`.
+#' @param ef_metric Character string specifying the efficiency metric:
+#'   \itemize{
+#'     \item `"Pace_HR"`: Pace-based efficiency (for running). Formula: (distance / moving_time) / avg_HR
+#'     \item `"Power_HR"`: Power-based efficiency (for cycling). Formula: avg_watts / avg_HR
+#'   }
+#'   Default: `c("Pace_HR", "Power_HR")` (uses first matching metric for activity type).
+#' @param start_date Optional. Analysis start date (YYYY-MM-DD string, Date, or POSIXct).
+#'   Defaults to one year before `end_date`.
+#' @param end_date Optional. Analysis end date (YYYY-MM-DD string, Date, or POSIXct).
+#'   Defaults to current date (Sys.Date()).
+#' @param min_duration_mins Numeric. Minimum activity duration in minutes to include
+#'   in analysis (default: 20). Filters out very short activities that may not
+#'   represent steady-state aerobic efforts.
 #'
-#' @return A data frame with columns: date, activity_type, ef_value.
+#' @return A tibble with the following columns:
+#' \describe{
[…]
