+++ R/calculate_decoupling.R
@@ -1,33 +1,116 @@
 # R/calculate_decoupling.R
 
-#' Calculate Aerobic Decoupling from Local Strava Data
+#' Calculate Cardiovascular Decoupling
 #'
-#' Calculates cardiovascular (aerobic) decoupling by analyzing the relationship
-#' between normalized power/pace and heart rate. Decoupling indicates cardiovascular
-#' drift, where heart rate rises relative to power/pace output during sustained efforts.
+#' Quantifies cardiovascular drift during endurance activities by comparing the
+#' efficiency (pace/HR or power/HR) between the first and second halves of an activity.
+#' Decoupling is a key indicator of aerobic endurance and durability (Maunder et al., 2021).
+#'
+#' @description
+#' Cardiovascular decoupling (aerobic decoupling) measures the deterioration of
+#' efficiency during prolonged exercise. As you fatigue, heart rate tends to rise
+#' (cardiac drift) while power/pace may decrease, indicating reduced efficiency.
+#' Low decoupling suggests good aerobic fitness and durability at the given intensity.
+#'
+#' **Decoupling Formula:**
+#' 
+#' Decoupling\% = ((EF_half1 - EF_half2) / EF_half1) × 100
+#' 
+#' Where:
+#' \itemize{
+#'   \item **For Running**: EF = Pace / Heart Rate
+#'   \item **For Cycling**: EF = Power / Heart Rate
+#' }
+#'
+#' **Interpretation:**
+#' \itemize{
+#'   \item **< 5\%**: Excellent aerobic fitness and pacing
+#'   \item **5-10\%**: Acceptable for the intensity
+#'   \item **> 10\%**: Possible inadequate aerobic base, too fast pace, or environmental stress
+#'   \item **Negative values**: Possible warm-up effect or pacing issues
+#' }
 #'
 #' @param activities_data A data frame of activities from `load_local_activities()`.
-#'   Must contain columns: date, type, filename.
-#' @param export_dir Base directory of the Strava export containing the activities folder.
-#'   Default is "strava_export_data".
-#' @param activity_type Type of activities to analyze (e.g., "Run", "Ride").
-#' @param start_date Optional start date for analysis (YYYY-MM-DD). Defaults to NULL (all dates).
-#' @param end_date Optional end date for analysis (YYYY-MM-DD). Defaults to NULL (all dates).
-#' @param min_duration_mins Minimum activity duration in minutes to include. Default 40.
+#'   Must contain columns: `date`, `type`, `filename`. The `filename` column is used
+#'   to locate detailed activity files (FIT/TCX/GPX) in the Strava export.
+#' @param export_dir Character string. Base directory of the Strava export containing
+#'   the activities folder with detailed activity files. Default: `"strava_export_data"`.
+#'   If you loaded from a ZIP, this should be the path where the ZIP was extracted
+#'   (handled automatically by `load_local_activities`).
+#' @param activity_type Character vector or single string. Activity type(s) to analyze.
+#'   Common values: `"Run"`, `"Ride"`. Default: `c("Run", "Ride")`.
+#' @param start_date Optional. Analysis start date (YYYY-MM-DD string, Date, or POSIXct).
+#'   Defaults to NULL (include all dates).
+#' @param end_date Optional. Analysis end date (YYYY-MM-DD string, Date, or POSIXct).
+#'   Defaults to NULL (include all dates).
+#' @param min_duration_mins Numeric. Minimum activity duration in minutes to include
+#'   in analysis (default: 40). Shorter activities may not show meaningful decoupling.
+#'   Recommended: 45-60+ minutes for reliable decoupling assessment.
 #'
[…]
