+++ man/athlytics_palette_nature.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/color_palettes.R
+\name{athlytics_palette_nature}
+\alias{athlytics_palette_nature}
+\title{Nature-Inspired Color Palette}
+\usage{
+athlytics_palette_nature()
+}
+\description{
+Professional, colorblind-friendly palette based on Nature journal's
+visualization guidelines. Suitable for multi-series plots.
+}
+\examples{
+\dontrun{
+library(ggplot2)
+ggplot(data, aes(x, y, color = group)) +
+  geom_line() +
+  scale_color_manual(values = athlytics_palette_nature())
+}
+}
