+++ man/plot_acwr.Rd
@@ -17,7 +17,9 @@ plot_acwr(
   user_resting_hr = NULL,
   smoothing_period = 7,
   highlight_zones = TRUE,
-  acwr_df = NULL
+  acwr_df = NULL,
+  group_var = NULL,
+  group_colors = NULL
 )
 }
 \arguments{
