+++ man/athlytics_palette_vibrant.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/color_palettes.R
+\name{athlytics_palette_vibrant}
+\alias{athlytics_palette_vibrant}
+\title{Vibrant High-Contrast Palette}
+\usage{
+athlytics_palette_vibrant()
+}
+\description{
+High-saturation palette optimized for presentations and posters.
+Maximum visual impact while maintaining colorblind accessibility.
+}
+\examples{
+\dontrun{
+library(ggplot2)
+ggplot(data, aes(x, y, fill = category)) +
+  geom_bar(stat = "identity") +
+  scale_fill_manual(values = athlytics_palette_vibrant())
+}
+}
