### 2025-10-10 6a4f2f4
Update core functions and documentation
