+++ R/plot_pbs.R
@@ -2,24 +2,21 @@
 #'
 #' Visualizes the trend of personal best times for specific running distances.
 #'
-#' Plots the trend of best efforts for specified distances, highlighting new PBs.
-#' Uses pre-calculated data or calls `calculate_pbs`.
+#' Plots the trend of best efforts for specified distances, highlighting new PBs
+#' using pre-calculated data from `calculate_pbs`.
 #'
-#' @param stoken A valid Strava token from `rStrava::strava_oauth()`. Required unless `pbs_df` is provided.
-#' @param activity_type Type(s) of activities to search (e.g., "Run"). Default "Run".
-#' @param distance_meters Numeric vector of distances (meters) to plot PBs for (e.g., `c(1000, 5000)`).
-#'   Relies on Strava's `best_efforts` data.
-#' @param max_activities Max number of recent activities to check. Default 500. Reduce for speed.
-#' @param date_range Optional. Filter activities by date `c("YYYY-MM-DD", "YYYY-MM-DD")`.
+#' @param pbs_df A data frame from `calculate_pbs` containing PB data.
+#'   Must contain columns: 'activity_date', 'time_seconds', 'distance_label', 'is_pb'.
+#' @param activity_type Type(s) of activities to display (e.g., "Run"). Default "Run".
+#' @param distance_meters Optional. Numeric vector of distances (meters) to filter.
+#'   If NULL, all distances in pbs_df are plotted.
 #' @param add_trend_line Logical. Whether to add a trend line to the plot. Default TRUE.
-#' @param pbs_df Optional. A pre-calculated data frame from `calculate_pbs`.
-#'   If provided, `stoken` and other calculation parameters are ignored.
 #'
 #' @return A ggplot object showing PB trends, faceted by distance if multiple are plotted.
 #'
 #' @details Visualizes data from `calculate_pbs`. Points show best efforts;
 #'   solid points mark new PBs. Y-axis is MM:SS.
-#'   If `pbs_df` is not provided, calls `calculate_pbs` first (can be slow).
+#'   Users must first call `calculate_pbs` to generate the required data.
 #'
 #' 
 #' @importFrom dplyr filter select mutate arrange group_by slice bind_rows summarise distinct rename %>% left_join
@@ -32,115 +29,38 @@
 #' @export
 #'
 #' @examples
-#' # Example using simulated data
-#' data(athlytics_sample_data)
-#' # athlytics_sample_pbs should contain the PBs to be plotted
-#' if (!is.null(athlytics_sample_pbs) && nrow(athlytics_sample_pbs) > 0) {
-#'   sample_pbs_for_plot <- athlytics_sample_pbs
-#'   
-  #'   # Note: date column should be named 'activity_date' and be of Date type
-#'   if ("date" %in% names(sample_pbs_for_plot) && !"activity_date" %in% names(sample_pbs_for_plot)) {
-#'     names(sample_pbs_for_plot)[names(sample_pbs_for_plot) == "date"] <- "activity_date"
-#'   }
-#'   if ("activity_date" %in% names(sample_pbs_for_plot)) {
-#'     sample_pbs_for_plot$activity_date <- as.Date(sample_pbs_for_plot$activity_date)
-#'   } else {
-#'     message("Relevant date column not found in sample PBs for example.")
-#'   }
-#'   
-#'   # plot_pbs requires distance_meters. Extract from sample data.
-#'   req_dist_meters <- NULL
-#'   if ("distance" %in% names(sample_pbs_for_plot)) {
-#'     req_dist_meters <- unique(sample_pbs_for_plot$distance)
-#'   } else if ("distance_target_m" %in% names(sample_pbs_for_plot)) {
[…]
