+++ R/plot_decoupling.R
@@ -39,6 +39,8 @@ explicit_english_month_year <- function(date_obj) {
 #' @param decoupling_df Optional. A pre-calculated data frame from `calculate_decoupling`.
 #'   If provided, `stoken` and other calculation parameters are ignored.
 #'   Must contain 'date' and 'decoupling' columns.
+#' @param group_var Optional. Column name for grouping/faceting (e.g., "athlete_id").
+#' @param group_colors Optional. Named vector of colors for groups.
 #'
 #' @return A ggplot object showing the decoupling trend.
 #'
@@ -62,13 +64,10 @@ explicit_english_month_year <- function(date_obj) {
 #' @export
 #'
 #' @examples
-#' # Example using simulated data
-#' data(athlytics_sample_data)
-#' # Explicitly name decoupling_df and provide activity_type
-#' if (!is.null(athlytics_sample_decoupling) && nrow(athlytics_sample_decoupling) > 0) {
-#'   p <- plot_decoupling(decoupling_df = athlytics_sample_decoupling, activity_type = "Run")
-#'   print(p)
-#' }
+#' # Example using pre-calculated sample data
+#' data("athlytics_sample_decoupling", package = "Athlytics")
+#' p <- plot_decoupling(athlytics_sample_decoupling)
+#' print(p)
 #'
 #' \dontrun{
 #' # Example using real data (requires authentication)
