### 2025-10-10 e314a5f
Add pkgdown config to organize documentation structure
