### 2025-10-10 e6c1a7e
Add GitHub Actions workflow for pkgdown documentation
