### 2025-10-10 2675e23
Remove redundant LICENSE.md file
