### 2025-10-10 e6786e5
Fix documentation issues: Update NEWS.md to v1.0.0, fix encoding and heading levels, reorganize images to man/figures, remove Athlytics-introduction vignette
