### 2025-10-10 fcf7902
Fix LICENSE file with complete MIT license text
