### 2025-10-10 1edc38a
Remove duplicate LICENSE.md file to fix double license display
