+++ tests/testthat/test-calculate_pbs.R
@@ -1,250 +1,0 @@
-# tests/testthat/test-calculate_pbs.R
-
-library(testthat)
-library(Athlytics)
-library(mockery)
-# library(rStrava) # Not calling rStrava directly
-
-# Load data: sample data from package & mock API returns from helper
-data(athlytics_sample_data)
-source(test_path("helper-mockdata.R"), local = TRUE)
-
-# Mock Strava token
-mock_stoken <- structure(list(token = list(access_token = "fake_token")), class = "Token2.0")
-
-# --- Test Parameter Validation ---
-
-test_that("calculate_pbs throws error for non-Token2.0 stoken", {
-  expect_error(
-    calculate_pbs(stoken = data.frame(), distance_meters = 1000),
-    # The error message might change with proper mocking if it gets past stoken check
-    # Assumes stoken type check first
-    regexp = "Assertion on 'stoken' failed: Must inherit from class 'Token2.0', but has class 'data.frame'\\."
-  )
-})
-
-test_that("calculate_pbs throws error for non-numeric distance_meters", {
-  expect_error(
-    calculate_pbs(stoken = mock_stoken, distance_meters = "1km"),
-    regexp = "'distance_meters' must be a numeric vector of distances \\(e\\.g\\., c\\(1000, 5000\\)\\)\\."
-  )
-})
-
-test_that("calculate_pbs throws error for zero or negative distance_meters", {
-    expect_error(
-        calculate_pbs(stoken = mock_stoken, distance_meters = c(1000, 0)),
-        regexp = "All 'distance_meters' must be positive\\."
-    )
-    expect_error(
-        calculate_pbs(stoken = mock_stoken, distance_meters = -1000),
-        regexp = "All 'distance_meters' must be positive\\."
-    )
-})
-
-# --- Test API Interaction and Core Logic with Mocking ---
-
-# Helper to create a mock rStrava::get_efforts_list return value
-# This function will simulate finding best efforts for given distances within a single activity's efforts list
-# It's a simplified mock focusing on returning plausible data for calculate_pbs
-
-# The actual get_efforts_list returns a list of data frames, one for each segment effort in an activity.
-# Each data frame has columns like: id, resource_state, name, activity, athlete, elapsed_time, moving_time, start_date, start_date_local, distance, start_index, end_index, pr_rank, segment.id, segment.name, etc.
-# For calculate_pbs, we are interested in finding the best `elapsed_time` for a `distance` that is close to the target distances.
-# The function internally seems to look for `segment.distance` matching `distance_meters`.
-  # Mock returns list of efforts for one activity
-# To test calculate_pbs properly, we need to simulate get_efforts_list being called for *each* activity from get_activity_list
-
-# Mock for get_efforts_list: needs to be adaptable or we mock its caller if calculate_pbs has an internal helper.
-# calculate_pbs directly calls get_efforts_list(act_data = activity_list, stoken = stoken, id = current_activity_id)
-# So, the mock needs to handle different `id` arguments if we want to simulate different efforts for different activities.
-# For simplicity, let's make it return the *same* set of efforts for any activity ID, but this means PBs will always come from the same mocked efforts.
-
-# Adjusted mock_efforts_data_for_activity to be a list of lists
-mock_efforts_data_for_activity <- list(
-  list( # Effort 1
-    name = "Sprint to Tree",
-    distance = 100, # meters
-    elapsed_time = 15, # seconds
-    start_date_local = "2023-10-01T10:00:00Z" # Using start_date_local if that's what rStrava provides
-  ),
-  list( # Effort 2
-    name = "400m Dash",
-    distance = 400, # meters
[…]
