### 2025-10-10 e24e9d4
Restore correct versions of decoupling, exposure and pbs functions to pass all tests
