+++ R/plot_decoupling.R
@@ -39,8 +39,6 @@ explicit_english_month_year <- function(date_obj) {
 #' @param decoupling_df Optional. A pre-calculated data frame from `calculate_decoupling`.
 #'   If provided, `stoken` and other calculation parameters are ignored.
 #'   Must contain 'date' and 'decoupling' columns.
-#' @param group_var Optional. Column name for grouping/faceting (e.g., "athlete_id").
-#' @param group_colors Optional. Named vector of colors for groups.
 #'
 #' @return A ggplot object showing the decoupling trend.
 #'
@@ -49,7 +47,7 @@ explicit_english_month_year <- function(date_obj) {
 #'   used as reference. If `decoupling_df` is not provided, calls `calculate_decoupling` first
 #'   (can be slow and hit API limits).
 #'
-#' 
+#' @importFrom rStrava get_activity_list get_activity_streams
 #' @importFrom dplyr filter select mutate arrange %>% rename left_join case_when group_by summarise pull first last tibble
 #' @importFrom lubridate as_date
 #' @importFrom lubridate date
@@ -64,12 +62,15 @@ explicit_english_month_year <- function(date_obj) {
 #' @export
 #'
 #' @examples
-#' # Example using pre-calculated sample data
-#' data("athlytics_sample_decoupling", package = "Athlytics")
-#' p <- plot_decoupling(athlytics_sample_decoupling)
-#' print(p)
+#' # Example using simulated data
+#' data(Athlytics_sample_data)
+#' # Explicitly name decoupling_df and provide activity_type
+#' if (!is.null(athlytics_sample_decoupling) && nrow(athlytics_sample_decoupling) > 0) {
+#'   p <- plot_decoupling(decoupling_df = athlytics_sample_decoupling, activity_type = "Run")
+#'   print(p)
+#' }
 #'
-#' \dontrun{
+#' \donttest{
 #' # Example using real data (requires authentication)
 #' # NOTE: The following rStrava::strava_oauth call is a placeholder.
 #' # You MUST replace placeholders with your actual Strava API credentials.
@@ -83,7 +84,7 @@ explicit_english_month_year <- function(date_obj) {
 #'   )
 #'
 #'   if (inherits(stoken_example, "Token2.0")) {
-#'     message("Placeholder stoken_example created for athlytics examples.")
+#'     message("Placeholder stoken_example created for Athlytics examples.")
 #'
 #'     # Example 1: Plot Decoupling trend for Runs (last 6 months)
 #'     # This first calculates the data, then plots it.
@@ -157,32 +158,27 @@ plot_decoupling <- function(stoken,
                             max_activities = 50,
                             add_trend_line = TRUE,
                             smoothing_method = "loess",
-                            decoupling_df = NULL,
-                            group_var = NULL,
-                            group_colors = NULL) {
+                            decoupling_df = NULL) {
 
   # Match arg for decouple_metric to ensure only one is used internally if multiple are provided
   decouple_metric_label <- match.arg(decouple_metric)
 
-  # --- Check if first argument is already decoupling data frame ---
-  if (is.data.frame(stoken) && any(c("decoupling_pct", "decoupling_percent", "decoupling") %in% colnames(stoken))) {
-    decoupling_df <- stoken
-  }
-
   # --- Get Data --- 
   # If decoupling_df is not provided, calculate it
   if (is.null(decoupling_df)) {
-      # Check if stoken provided when decoupling_df is not
[…]
