### 2025-10-10 8e23118
Add vignettes for athlytics introduction and advanced features
