+++ vignettes/advanced_features.Rmd
@@ -1,0 +1,406 @@
+---
+title: "Advanced Features: EWMA, Quality Control, and Cohort Analysis"
+author: "Athlytics Package"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Advanced Features: EWMA, Quality Control, and Cohort Analysis}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7,
+  fig.height = 5
+)
+```
+
+# Introduction
+
+This vignette demonstrates the advanced features of Athlytics for research-grade sports physiology analysis:
+
+1. **Data Quality Control**: Detecting and flagging HR/power spikes, GPS drift
+2. **ACWR with EWMA**: Exponentially weighted moving average with confidence bands
+3. **Cohort Reference Analysis**: Multi-athlete percentile comparisons
+
+These features make Athlytics suitable for both individual training analysis and multi-athlete research studies.
+
+---
+
+# 1. Data Quality Control
+
+Before calculating physiological metrics, it's crucial to ensure data quality. The `flag_quality()` function detects common issues in activity stream data.
+
+## Basic Quality Checking
+
+```{r eval=FALSE}
+library(Athlytics)
+library(dplyr)
+
+# Load a single activity with stream data
+activities <- load_local_activities("export_12345678.zip")
+
+# Parse stream data from a FIT/TCX file
+# (Assuming you have stream data - this is a conceptual example)
+stream_data <- parse_activity_file("path/to/activity.fit")
+
+# Flag quality issues
+flagged_data <- flag_quality(
+  stream_data,
+  sport = "Run",
+  hr_range = c(30, 220),           # Valid HR range
+  pw_range = c(0, 1500),            # Valid power range
+  max_run_speed = 7.0,              # Max speed (m/s) ≈ 2:23/km pace
+  max_hr_jump = 10,                 # Max HR change per second
+  max_pw_jump = 300,                # Max power change per second
+  min_steady_minutes = 20,          # Min duration for steady-state
+  steady_cv_threshold = 8           # CV% threshold for steady-state
+)
+
+# View quality summary
+summary_stats <- quality_summary(flagged_data)
+print(summary_stats)
+```
+
+**Output interpretation:**
+- `quality_score`: Overall quality (0-1), where 1 = perfect
+- `flagged_pct`: Percentage of data points with issues
+- `steady_state_pct`: Percentage suitable for EF/decoupling calculations
+
+## Why Quality Control Matters
+
+**Without quality control:**
+- HR spikes from sensor errors inflate HRSS calculations
+- GPS drift creates artificially high speeds, affecting pace metrics
+- Non-steady-state segments contaminate Efficiency Factor (EF) calculations
+
+**With quality control:**
+- Only clean data contributes to metrics
+- Steady-state detection ensures valid EF/decoupling calculations
+- Researchers can report data quality in publications
+
+---
+
+# 2. ACWR with EWMA Method
+
+Traditional ACWR uses rolling averages (RA). The EWMA method offers:
+- **Greater sensitivity** to recent training changes
+- **Customizable decay rates** via half-life parameters
+- **Confidence bands** via bootstrap to quantify uncertainty
+
+## Comparing RA vs EWMA
+
[…]
