### 2025-10-10 c36612f
Add vignette HTML documentation to inst/doc folder
