+++ inst/doc/advanced_features.R
@@ -1,0 +1,210 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7,
+  fig.height = 5
+)
+
+## ----eval=FALSE---------------------------------------------------------------
+# library(Athlytics)
+# library(dplyr)
+# 
+# # Load a single activity with stream data
+# activities <- load_local_activities("export_12345678.zip")
+# 
+# # Parse stream data from a FIT/TCX file
+# # (Assuming you have stream data - this is a conceptual example)
+# stream_data <- parse_activity_file("path/to/activity.fit")
+# 
+# # Flag quality issues
+# flagged_data <- flag_quality(
+#   stream_data,
+#   sport = "Run",
+#   hr_range = c(30, 220),           # Valid HR range
+#   pw_range = c(0, 1500),            # Valid power range
+#   max_run_speed = 7.0,              # Max speed (m/s) ≈ 2:23/km pace
+#   max_hr_jump = 10,                 # Max HR change per second
+#   max_pw_jump = 300,                # Max power change per second
+#   min_steady_minutes = 20,          # Min duration for steady-state
+#   steady_cv_threshold = 8           # CV% threshold for steady-state
+# )
+# 
+# # View quality summary
+# summary_stats <- quality_summary(flagged_data)
+# print(summary_stats)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Load activities
+# activities <- load_local_activities("export_12345678.zip")
+# 
+# # Calculate ACWR using Rolling Average (traditional)
+# acwr_ra <- calculate_acwr_ewma(
+#   activities,
+#   activity_type = "Run",
+#   method = "ra",
+#   acute_period = 7,
+#   chronic_period = 28,
+#   load_metric = "duration_mins"
+# )
+# 
+# # Calculate ACWR using EWMA
+# acwr_ewma <- calculate_acwr_ewma(
+#   activities,
+#   activity_type = "Run",
+#   method = "ewma",
+#   half_life_acute = 3.5,    # Acute load half-life (days)
+#   half_life_chronic = 14,   # Chronic load half-life (days)
+#   load_metric = "duration_mins"
+# )
+# 
+# # Compare the two methods
+# library(ggplot2)
+# ggplot() +
+#   geom_line(data = acwr_ra, aes(x = date, y = acwr_smooth),
+#             color = "blue", size = 1, linetype = "solid") +
+#   geom_line(data = acwr_ewma, aes(x = date, y = acwr_smooth),
+#             color = "red", size = 1, linetype = "dashed") +
+#   labs(title = "ACWR: Rolling Average vs EWMA",
+#        subtitle = "Blue = RA (7:28d) | Red = EWMA (3.5:14d half-life)",
+#        x = "Date", y = "ACWR") +
+#   theme_minimal()
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Calculate EWMA with 95% confidence bands
+# acwr_ewma_ci <- calculate_acwr_ewma(
+#   activities,
+#   activity_type = "Run",
+#   method = "ewma",
+#   half_life_acute = 3.5,
+#   half_life_chronic = 14,
+#   load_metric = "duration_mins",
+#   ci = TRUE,                # Enable confidence intervals
+#   B = 200,                  # Bootstrap iterations
+#   block_len = 7,            # Block length (days) for bootstrap
+#   conf_level = 0.95         # 95% CI
+# )
+# 
+# # Plot with confidence bands
[…]

[…]
