+++ pkgdown.yml
@@ -1,0 +1,10 @@
+pandoc: 3.1.11
+pkgdown: 2.1.3
+pkgdown_sha: ~
+articles:
+  advanced_features: advanced_features.html
+  athlytics_introduction: athlytics_introduction.html
+last_built: 2025-10-10T19:21Z
+urls:
+  reference: https://hezhiang.com/Athlytics/reference
+  article: https://hezhiang.com/Athlytics/articles
