### 2025-10-10 2ad5701
Refactor to local-first architecture and improve API consistency
Major changes:
- Remove all rStrava dependencies from NAMESPACE and plot functions
- Export calculate_acwr_ewma and flag_quality functions
- Refactor calculate_decoupling to use local FIT/TCX/GPX parsing
- Rename load_local_activities parameter: csv_path -> path
- Update README with unified metric naming (pace_hr, power_hr)
- Update README multi-athlete example to use group_modify()
- Fix documentation: escape percentage signs, add markdown support
- Update all plot_* functions to emphasize local data over API
- Remove deprecated
[…]
