### 2025-10-10 84529ec
Fix README: translate Chinese comments to English
