### 2025-10-10 b341806
Enhance scientific rigor: unify parameter naming and improve documentation
Scientific improvements:
- Unify metric naming: pace_hr/power_hr (backward compatible with Pace_HR/Power_HR)
- Add clear unit specifications: m路s鈦宦孤穊pm鈦宦?for pace_hr, W路bpm鈦宦?for power_hr
- Enhance ACWR documentation: emphasize daily aggregation and zero-fill logic
- Improve EWMA documentation: clarify half-life formula and bootstrap methodology
- Update all examples to use consistent lowercase naming
- Add status field documentation for EF/decoupling quality reporting
- All tests passing: 146 PASS, 0 FAIL
