+++ R/calculate_acwr_ewma.R
@@ -35,12 +35,15 @@
 #'     with configurable half-lives. More responsive to recent changes.
 #' }
 #'
-#' For EWMA, the smoothing parameter alpha is calculated as: alpha = ln(2) / half_life
+#' **EWMA Formula**: The smoothing parameter α is calculated from half-life: 
+#' `α = ln(2) / half_life`. The EWMA update is: `E_t = α × L_t + (1-α) × E_{t-1}`
+#' where L_t is daily load and E_t is the exponentially weighted average.
 #'
-#' **Confidence Bands**: When `ci = TRUE` and `method = "ewma"`, uses moving-block
-#' bootstrap to estimate uncertainty. The daily load sequence is resampled in
+#' **Confidence Bands**: When `ci = TRUE` and `method = "ewma"`, uses **moving-block
+#' bootstrap** to estimate uncertainty. The daily load sequence is resampled in
 #' weekly blocks (preserving within-week correlation), ACWR is recalculated,
-#' and percentiles form the confidence bands.
+#' and percentiles form the confidence bands. This accounts for temporal correlation
+#' in training load patterns.
 #'
 #' @importFrom dplyr filter select mutate group_by summarise arrange %>% left_join coalesce
 #' @importFrom lubridate as_date days
