### 2025-10-10 540f534
Fix article links in README to use correct file naming
- Fix introduction link: Athlytics-introduction.html 鈫?athlytics_introduction.html
- Fix advanced features link: advanced-features.html 鈫?advanced_features.html
- Use underscore naming convention to match actual pkgdown generated files
