### 2025-10-10 f37db80
Fix critical P0 scientific issues
- Fix activity_type to be required (prevent mixing incompatible sports)
- Clarify EF calculation uses speed (m/s) not pace
- Keep highlight_zones = TRUE as requested
- All other P0 issues need further investigation:
  * EF/Decoupling missing steady-state gating
  * HR coverage thresholds not implemented
  * Quality control not internal to calculation functions
