### 2025-10-10 44335d7
Add steady-state gating to EF and Decoupling calculations
Critical P0 fixes:
- Add min_steady_minutes and steady_cv_threshold parameters to calculate_ef()
- Add min_steady_minutes and steady_cv_threshold parameters to calculate_decoupling()
- Add status field to EF results: 'ok', 'too_short', 'insufficient_data'
- Add status field to Decoupling results: 'ok', 'non_steady', 'insufficient_data'
- Enforce minimum duration requirements for steady-state analysis
- Add parameter validation for steady-state thresholds

This addresses the critical issue where EF/Decoupling were calculated
without stea
[…]
