### 2025-10-10 049ce88
Add HR coverage validation and TSS/HRSS parameter validation
Critical P0 fixes:
- Add min_hr_coverage parameter to calculate_ef() and calculate_decoupling()
- Add HR coverage validation (simplified version using average_hr)
- Add status field 'insufficient_hr_data' for low HR coverage activities
- Add TSS/HRSS parameter validation to calculate_acwr_ewma()
- Validate user_ftp required for TSS, user_max_hr/user_resting_hr for HRSS
- Add load_metric validation with proper error messages

This addresses critical issues where:
1. EF/Decoupling were calculated without HR coverage validation
2. TSS/H
[…]
