### 2025-10-10 704e416
Add quality control integration to EF and Decoupling calculations
P0 fixes:
- Add quality_control parameter to calculate_ef() and calculate_decoupling()
- Support 'off', 'flag', and 'filter' modes for data quality handling
- Default to 'filter' mode for scientific rigor
- Add parameter validation for quality_control options
- Prepare foundation for integrating flag_quality() function

This addresses the critical issue where quality control was not
integrated into calculation functions, allowing potentially
unreliable data to be used in physiological metrics.
