### 2025-10-10 f72efc5
Update documentation and pkgdown site with P0 scientific fixes
- Regenerate Rd files with new parameters (quality_control, min_hr_coverage, etc.)
- Update pkgdown site with latest function documentation
- Include all P0 scientific rigor improvements in online documentation
- Ensure website reflects steady-state gating, HR coverage validation, and parameter validation
