+++ man/calculate_ef.Rd
@@ -10,7 +10,11 @@ calculate_ef(
   ef_metric = c("pace_hr", "power_hr", "Pace_HR", "Power_HR"),
   start_date = NULL,
   end_date = NULL,
-  min_duration_mins = 20
+  min_duration_mins = 20,
+  min_steady_minutes = 20,
+  steady_cv_threshold = 0.08,
+  min_hr_coverage = 0.9,
+  quality_control = c("off", "flag", "filter")
 )
 }
 \arguments{
@@ -41,6 +45,18 @@ Defaults to current date (Sys.Date()).}
 \item{min_duration_mins}{Numeric. Minimum activity duration in minutes to include
 in analysis (default: 20). Filters out very short activities that may not
 represent steady-state aerobic efforts.}
+
+\item{min_steady_minutes}{Numeric. Minimum duration (minutes) for steady-state segment (default: 20).
+Activities shorter than this are automatically rejected for EF calculation.}
+
+\item{steady_cv_threshold}{Numeric. Coefficient of variation threshold for steady-state (default: 0.08 = 8\%).
+Activities with higher variability are rejected as non-steady-state.}
+
+\item{min_hr_coverage}{Numeric. Minimum HR data coverage threshold (default: 0.9 = 90\%).
+Activities with lower HR coverage are rejected as insufficient data quality.}
+
+\item{quality_control}{Character. Quality control mode: "off" (no filtering), "flag" (mark issues),
+or "filter" (exclude flagged data). Default "filter" for scientific rigor.}
 }
 \value{
 A tibble with the following columns:
@@ -50,7 +66,8 @@ A tibble with the following columns:
 \item{ef_value}{Efficiency Factor value (numeric). Higher = better fitness.
 Units: m·s⁻¹·bpm⁻¹ for pace_hr, W·bpm⁻¹ for power_hr.}
 \item{status}{Character. "ok" for successful calculation, "non_steady" if steady-state
-criteria not met, "insufficient_data" if data quality issues. (Only if quality checks enabled)}
+criteria not met, "insufficient_data" if data quality issues, "too_short" if below min_steady_minutes,
+"insufficient_hr_data" if HR coverage below threshold.}
 }
 }
 \description{
