### 2025-10-10 a48b751
Fix critical P0 scientific rigor issues
Critical fixes:
1. Force explicit activity_type specification in ACWR calculations
   - Prevents mixing incompatible sports (Run+Ride) with different load metrics
   - Both calculate_acwr() and calculate_acwr_ewma() now require activity_type
   - Clear error message explains why mixing is problematic

2. Add quality control framework to EF calculations
   - Integrated quality_control parameter with 'off', 'flag', 'filter' modes
   - Added basic HR quality gates (50-220 bpm range)
   - Foundation for full flag_quality() integration

3. Clarify HRSS/TSS im
[…]
