### 2025-10-10 3db4a25
Complete P0 scientific rigor fixes and add sample data
Final P0 fixes:
1. Add quality control to decoupling calculations
   - Integrated quality_control parameter in calculate_single_decoupling()
   - Added basic quality gates for velocity, power, and HR ranges
   - Foundation for full flag_quality() integration in stream processing

2. Create comprehensive sample data files
   - athlytics_sample_acwr.rda: 365 days of realistic ACWR progression
   - athlytics_sample_decoupling.rda: 365 days of decoupling percentages
   - athlytics_sample_ef.rda: 50 activities with EF values
   - athlytics_samp
[…]
