### 2025-10-10 1d74e9b
Fix sample data examples in documentation
- Fixed calculate_ef.R example to use athlytics_sample_ef instead of non-existent athlytics_sample_data
- Fixed calculate_acwr.R example to use athlytics_sample_acwr instead of non-existent athlytics_sample_data
- Fixed calculate_exposure.R example to use athlytics_sample_exposure instead of non-existent athlytics_sample_data
- Regenerated documentation and pkgdown site
- All sample datasets now load correctly without warnings
- HR plotting functions confirmed working with local data frames
