+++ R/calculate_ef.R
@@ -123,7 +123,7 @@
 #'
 #' @examples
 #' # Example using simulated data
-#' data(athlytics_sample_data)
+#' data(athlytics_sample_ef)
 #' print(head(athlytics_sample_ef))
 #'
 #' \dontrun{
