+++ R/calculate_exposure.R
@@ -35,7 +35,7 @@
 #'
 #' @examples
 #' # Example using simulated data
-#' data(athlytics_sample_data)
+#' data(athlytics_sample_exposure)
 #' print(head(athlytics_sample_exposure))
 #'
 #' \dontrun{
