### 2025-10-10 f2850ee
Fix documentation consistency issues for JOSS readiness
- Fix EF enum consistency: Replace all Pace_HR/Power_HR with pace_hr/power_hr across vignettes and R files
- Fix README ACWR example: Add activity_type='Run' to avoid cross-sport mixing for scientific accuracy
- Fix sample data documentation: Replace X rows placeholders with actual dimensions (365/50/100 rows)
- Fix NEWS outdated syntax: Replace group_by/do with modern group_modify()
- Update pkgdown site with all documentation improvements

All changes improve scientific rigor and documentation consistency for JOSS submission.
