+++ R/plot_decoupling.R
@@ -29,7 +29,7 @@ explicit_english_month_year <- function(date_obj) {
 #' @param stoken **Recommended: Pass pre-calculated data via `decoupling_df` (local export preferred).**
 #'   For legacy API usage: A Strava token from `rStrava::strava_oauth()`. This parameter is deprecated.
 #' @param activity_type Type(s) of activities to analyze (e.g., "Run", "Ride").
-#' @param decouple_metric Metric basis: "Pace_HR" or "Power_HR".
+#' @param decouple_metric Metric basis: "pace_hr" or "power_hr".
 #' @param start_date Optional. Analysis start date (YYYY-MM-DD string or Date). Defaults to ~1 year ago.
 #' @param end_date Optional. Analysis end date (YYYY-MM-DD string or Date). Defaults to today.
 #' @param min_duration_mins Minimum activity duration (minutes) to include. Default 45.
@@ -91,7 +91,7 @@ explicit_english_month_year <- function(date_obj) {
 #'         calculate_decoupling(
 #'             stoken = stoken_example,
 #'             activity_type = "Run",
-#'             decouple_metric = "Pace_HR",
+#'             decouple_metric = "pace_hr",
 #'             date_range = c(format(Sys.Date() - lubridate::months(6), "%Y-%m-%d"), 
 #'                            format(Sys.Date(), "%Y-%m-%d")),
 #'             max_activities = 5 # Keep low for example
@@ -112,7 +112,7 @@ explicit_english_month_year <- function(date_obj) {
 #'     # decoupling_rides <- calculate_decoupling(
 #'     #   stoken = stoken_example,
 #'     #   activity_type = "Ride",
-#'     #   decouple_metric = "Power_HR",
+#'     #   decouple_metric = "power_hr",
 #'     #   max_activities = 5
 #'     # )
 #'     # if (nrow(decoupling_rides) > 0 && "decoupling" %in% names(decoupling_rides)) {
@@ -126,7 +126,7 @@ explicit_english_month_year <- function(date_obj) {
 #'     # decoupling_multi_run <- calculate_decoupling(
 #'     #   stoken = stoken_example,
 #'     #   activity_type = c("Run", "VirtualRun"),
-#'     #   decouple_metric = "Pace_HR",
+#'     #   decouple_metric = "pace_hr",
 #'     #   max_activities = 5
 #'     # )
 #'     # if (nrow(decoupling_multi_run) > 0 && "decoupling" %in% names(decoupling_multi_run)) {
@@ -149,7 +149,7 @@ explicit_english_month_year <- function(date_obj) {
 #' }
 plot_decoupling <- function(stoken,
                             activity_type = c("Run", "Ride"), # Default to both if not specified
-                            decouple_metric = c("Pace_HR", "Power_HR"),
+                            decouple_metric = c("pace_hr", "power_hr"),
                             start_date = NULL,
                             end_date = NULL,
                             min_duration_mins = 45,
