+++ R/calculate_decoupling.R
@@ -197,21 +197,27 @@ calculate_decoupling <- function(activities_data = NULL,
     }
     
     # Calculate decoupling for this activity
-    decoupling_value <- tryCatch({
+    decoupling_result <- tryCatch({
       calculate_single_decoupling(stream_data, decouple_metric, quality_control)
     }, error = function(e) {
       message(sprintf("  Error calculating decoupling: %s", e$message))
-      return(NA_real_)
+      return(list(value = NA_real_, status = "calculation_error"))
     })
     
-    if (is.na(decoupling_value)) {
-      return(NULL)
+    # Handle both old format (just numeric) and new format (list with status)
+    if (is.list(decoupling_result)) {
+      decoupling_value <- decoupling_result$value
+      status <- decoupling_result$status
+    } else {
+      decoupling_value <- decoupling_result
+      status <- if (is.na(decoupling_value)) "insufficient_data" else "ok"
     }
     
-    # Return result
+    # Return result with status
     data.frame(
       date = activity$date,
       decoupling = decoupling_value,
+      status = status,
       stringsAsFactors = FALSE
     )
   })
@@ -253,7 +259,7 @@ calculate_single_decoupling <- function(stream_df, decouple_metric, quality_cont
     dplyr::filter(!is.na(.data$time), !is.na(.data$heartrate))
   
   if (nrow(stream_clean) < 100) {
-    stop("Insufficient data points for decoupling calculation (need at least 100 points).")
+    return(list(value = NA_real_, status = "insufficient_data_points"))
   }
   
   # Calculate velocity if needed
@@ -280,14 +286,11 @@ calculate_single_decoupling <- function(stream_df, decouple_metric, quality_cont
   }
   
   if (nrow(stream_clean) < 100) {
-    stop("Insufficient valid data after filtering (need at least 100 points with positive values).")
+    return(list(value = NA_real_, status = "insufficient_valid_data"))
   }
   
   # Apply quality control and steady-state gating
   if (quality_control != "off") {
-    # For now, we use simplified quality checks since we don't have full flag_quality integration
-    # In a full implementation, we would call flag_quality() here
-    
     # Basic quality gates
     if (decouple_metric == "pace_hr") {
       # Check for reasonable velocity values
@@ -304,17 +307,60 @@ calculate_single_decoupling <- function(stream_df, decouple_metric, quality_cont
       dplyr::filter(.data$heartrate > 50, .data$heartrate < 220)
     
     if (nrow(stream_clean) < 100) {
-      stop("Insufficient data after quality filtering (need at least 100 points).")
+      return(list(value = NA_real_, status = "insufficient_data_after_quality_filter"))
     }
   }
   
-  # Split into two halves
-  midpoint <- floor(nrow(stream_clean) / 2)
-  first_half <- stream_clean[1:midpoint, ]
-  second_half <- stream_clean[(midpoint + 1):nrow(stream_clean), ]
+  # Calculate HR coverage
+  hr_coverage <- sum(!is.na(stream_clean$heartrate) & stream_clean$heartrate > 0) / nrow(stream_clean)
+  if (hr_coverage < 0.9) {  # Default min_hr_coverage threshold
+    return(list(value = NA_real_, status = "insufficient_hr_data"))
+  }
   
-  # Calculate efficiency factor for each half
-  if (decouple_metric == "Pace_HR") {
+  # Find steady-state windows using rolling coefficient of variation
[…]
