### 2025-10-10 7ed8c47
Remove stoken parameter from all plot functions and update to data-first approach
- Replace stoken parameter with data parameter in all plot functions
- Update function signatures: plot_ef, plot_decoupling, plot_acwr, plot_exposure, plot_pbs
- Remove deprecated stoken parameter from calculate_decoupling
- Update all examples to use local Strava export data instead of API
- Improve documentation to emphasize local data workflow
- Maintain backward compatibility for direct data frame input
