+++ R/plot_decoupling.R
@@ -26,8 +26,8 @@ explicit_english_month_year <- function(date_obj) {
 #'
 #' Plots the aerobic decoupling trend over time. **Recommended workflow: Use local data via `decoupling_df`.**
 #'
-#' @param stoken **Recommended: Pass pre-calculated data via `decoupling_df` (local export preferred).**
-#'   For legacy API usage: A Strava token from `rStrava::strava_oauth()`. This parameter is deprecated.
+#' @param data **Recommended: Pass pre-calculated data via `decoupling_df` (local export preferred).**
+#'   A data frame from `calculate_decoupling()` or activities data from `load_local_activities()`.
 #' @param activity_type Type(s) of activities to analyze (e.g., "Run", "Ride").
 #' @param decouple_metric Metric basis: "pace_hr" or "power_hr".
 #' @param start_date Optional. Analysis start date (YYYY-MM-DD string or Date). Defaults to ~1 year ago.
@@ -38,14 +38,13 @@ explicit_english_month_year <- function(date_obj) {
 #' @param decoupling_df **Recommended.** A pre-calculated data frame from `calculate_decoupling()`.
 #'   When provided, analysis uses local data only (no API calls).
 #'   Must contain 'date' and 'decoupling' columns.
-#'
+
 #' @return A ggplot object showing the decoupling trend.
 #'
 #' @details Plots decoupling percentage ((EF_1st_half - EF_2nd_half) / EF_1st_half * 100).
 #'   Positive values mean HR drifted relative to output. A 5\\% threshold line is often
 #'   used as reference. **Best practice: Use `load_local_activities()` + `calculate_decoupling()` + this function.**
-#'   Legacy API mode is maintained for backward compatibility only.
-#'
+
 #' @importFrom dplyr filter select mutate arrange %>% rename left_join case_when group_by summarise pull first last tibble
 #' @importFrom lubridate as_date
 #' @importFrom lubridate date
@@ -53,101 +52,49 @@ explicit_english_month_year <- function(date_obj) {
 #' @importFrom lubridate ymd
 #' @importFrom lubridate ymd_hms
 #' @importFrom lubridate as_datetime
-#' @importFrom ggplot2 ggplot aes geom_point geom_smooth labs theme_minimal scale_x_date theme element_text scale_y_continuous annotate geom_hline theme_void ggtitle
+#' @importFrom ggplot2 ggplot aes geom_point geom_smooth labs theme_minimal scale_x_date theme element_text geom_hline
 #' @importFrom rlang .data
-#' @importFrom stats median na.omit
-#' @importFrom utils txtProgressBar setTxtProgressBar
 #' @export
 #'
 #' @examples
-#' # Example using simulated data
-#' data(Athlytics_sample_data)
-#' # Explicitly name decoupling_df and provide activity_type
-#' if (!is.null(athlytics_sample_decoupling) && nrow(athlytics_sample_decoupling) > 0) {
-#'   p <- plot_decoupling(decoupling_df = athlytics_sample_decoupling, activity_type = "Run")
-#'   print(p)
+#' # Example using pre-calculated sample data
+#' data("athlytics_sample_decoupling", package = "Athlytics")
+#' p <- plot_decoupling(athlytics_sample_decoupling)
+#' print(p)
+#'
+#' \dontrun{
+#' # Example using local Strava export data
+#' activities <- load_local_activities("strava_export_data/activities.csv")
+#' 
+#' # Example 1: Plot Decoupling trend for Runs (last 6 months)
+#' decoupling_runs_6mo <- calculate_decoupling(
+#'     activities_data = activities,
+#'     export_dir = "strava_export_data",
+#'     activity_type = "Run",
[…]

[…]
