### 2025-10-10 6daaae6
Fix hardcoded thresholds and unify enum naming
