### 2025-10-10 351962d
Clean up remaining stoken/API references and improve data quality handling
- Fix plot_exposure caption to use 'local Strava export' instead of 'rStrava'
- Correct plot_pbs example to use athlytics_sample_pbs data object
- Remove obsolete fetch_strava_activities.html from docs
- Improve calculate_ef() to return 'no_streams' status when using activity-level data
- Update documentation to clarify stream vs activity-level EF calculations
