+++ tests/testthat/test-decoupling.R
@@ -57,8 +57,8 @@ test_that("calculate_decoupling computes a plausible value from mock stream_df",
   # This test uses mock_activity_streams from helpe-mockdata.R for single activity stream processing
   expect_true(exists("mock_activity_streams"), "mock_activity_streams not loaded from helper.")
   
-  decoupling_power_hr <- calculate_decoupling(stream_df = mock_activity_streams, decouple_metric = "Power_HR")
-  decoupling_pace_hr <- calculate_decoupling(stream_df = mock_activity_streams, decouple_metric = "Pace_HR")
+  decoupling_power_hr <- calculate_decoupling(stream_df = mock_activity_streams, decouple_metric = "power_hr")
+  decoupling_pace_hr <- calculate_decoupling(stream_df = mock_activity_streams, decouple_metric = "pace_hr")
   
   expect_type(decoupling_power_hr, "double")
   expect_type(decoupling_pace_hr, "double")
@@ -72,19 +72,19 @@ test_that("calculate_decoupling computes a plausible value from mock stream_df",
 test_that("calculate_decoupling handles missing columns in stream_df", {
   expect_true(exists("mock_activity_streams"), "mock_activity_streams not loaded from helper.")
 
-  # Test missing watts for Power_HR
+  # Test missing watts for power_hr
   mock_streams_no_watts <- mock_activity_streams[, !(names(mock_activity_streams) %in% "watts")]
-  expect_error(calculate_decoupling(stream_df = mock_streams_no_watts, decouple_metric = "Power_HR"), 
+  expect_error(calculate_decoupling(stream_df = mock_streams_no_watts, decouple_metric = "power_hr"), 
                  regexp = "must contain 'watts' column")
                  
-  # Test missing velocity_smooth AND distance for Pace_HR
+  # Test missing velocity_smooth AND distance for pace_hr
   mock_streams_no_speed <- mock_activity_streams[, !(names(mock_activity_streams) %in% c("velocity_smooth", "distance"))]
-  expect_error(calculate_decoupling(stream_df = mock_streams_no_speed, decouple_metric = "Pace_HR"), 
+  expect_error(calculate_decoupling(stream_df = mock_streams_no_speed, decouple_metric = "pace_hr"), 
                  regexp = "must contain 'distance' or 'velocity_smooth' column")
                  
   # Test missing heartrate
   mock_streams_no_hr <- mock_activity_streams[, !(names(mock_activity_streams) %in% "heartrate")]
-  expect_error(calculate_decoupling(stream_df = mock_streams_no_hr, decouple_metric = "Power_HR"),
+  expect_error(calculate_decoupling(stream_df = mock_streams_no_hr, decouple_metric = "power_hr"),
                  regexp = "missing required columns")
 })
 
@@ -98,7 +98,7 @@ test_that("plot_decoupling returns a ggplot object using Athlytics_sample_data",
   
   # plot_decoupling requires decouple_metric for labeling if not directly derivable from decoupling_df
   # We assume athlytics_sample_decoupling might be for a mix or needs this specified.
-  p_decouple <- plot_decoupling(decoupling_df = athlytics_sample_decoupling, decouple_metric = "Pace_HR")
+  p_decouple <- plot_decoupling(decoupling_df = athlytics_sample_decoupling, decouple_metric = "pace_hr")
   expect_s3_class(p_decouple, "ggplot")
   
   # Check that the plot is not empty / uses the data
@@ -108,7 +108,7 @@ test_that("plot_decoupling returns a ggplot object using Athlytics_sample_data",
 
[…]
