+++ man/load_local_activities.Rd
@@ -87,7 +87,7 @@ acwr_data <- calculate_acwr(activities, load_metric = "distance_km")
 plot_acwr(acwr_data, highlight_zones = TRUE)
 
 # Multi-metric analysis
-ef_data <- calculate_ef(activities, ef_metric = "Pace_HR")
+ef_data <- calculate_ef(activities, ef_metric = "pace_hr")
 plot_ef(ef_data, add_trend_line = TRUE)
 }
 

[…]
