### 2025-10-10 f7f3ca6
Clean up documentation and improve consistency
