+++ man/plot_pbs.Rd
@@ -49,7 +49,7 @@ Legacy API mode is maintained for backward compatibility only.
 }
 \examples{
 # Example using simulated data
-data(Athlytics_sample_data)
+data(athlytics_sample_pbs)
 # athlytics_sample_pbs should contain the PBs to be plotted
 if (!is.null(athlytics_sample_pbs) && nrow(athlytics_sample_pbs) > 0) {
   sample_pbs_for_plot <- athlytics_sample_pbs
