### 2025-10-10 23d44dd
Replace deprecated do() with group_modify() in cohort examples
