+++ man/cohort_reference.Rd
@@ -73,7 +73,7 @@ cohort_data <- bind_rows(athlete1, athlete2, athlete3)
 # Calculate ACWR for each athlete
 cohort_acwr <- cohort_data \%>\%
   group_by(athlete_id) \%>\%
-  do(calculate_acwr_ewma(.))
+  group_modify(~calculate_acwr_ewma(.x))
 
 # Calculate reference percentiles
 reference <- cohort_reference(
