### 2025-10-10 4df8cb0
Fix tests: update for removed stoken and require activity_type
