### 2025-10-11 d2e4108
Add JOSS paper submission files (paper.md and paper.bib)
