+++ paper/paper.bib
@@ -1,0 +1,89 @@
+@article{gabbett2016,
+  author  = {Gabbett, Tim J.},
+  title   = {The training–injury prevention paradox: should athletes be training smarter and harder?},
+  journal = {British Journal of Sports Medicine},
+  year    = {2016},
+  volume  = {50},
+  number  = {5},
+  pages   = {273--280},
+  doi     = {10.1136/bjsports-2015-095788}
+}
+
+@article{impellizzeri2020acwr,
+  author  = {Impellizzeri, Franco M. and Tenan, Matthew S. and Kempton, Tom and Novak, Andrew and Coutts, Aaron J.},
+  title   = {Acute:Chronic Workload Ratio: Conceptual Issues and Fundamental Pitfalls},
+  journal = {International Journal of Sports Physiology and Performance},
+  year    = {2020},
+  volume  = {15},
+  number  = {6},
+  pages   = {907--913},
+  doi     = {10.1123/ijspp.2019-0864}
+}
+
+@article{kunsch1989,
+  author  = {K{\"u}nsch, Hans R.},
+  title   = {The Jackknife and the Bootstrap for General Stationary Observations},
+  journal = {The Annals of Statistics},
+  year    = {1989},
+  volume  = {17},
+  number  = {3},
+  pages   = {1217--1241},
+  doi     = {10.1214/aos/1176347265}
+}
+
+@article{politis1994,
+  author  = {Politis, Dimitris N. and Romano, Joseph P.},
+  title   = {The Stationary Bootstrap},
+  journal = {Journal of the American Statistical Association},
+  year    = {1994},
+  volume  = {89},
+  number  = {428},
+  pages   = {1303--1313},
+  doi     = {10.1080/01621459.1994.10476870}
+}
+
+@Manual{rStrava,
+  title   = {rStrava: Access the Strava API},
+  author  = {Marcus W. Beck and Pedro Villarroel and Daniel Padfield and Lorenzo Gaborini and Niklas von Maltzahn},
+  year    = {2024},
+  note    = {R package version 1.3.2},
+  doi     = {10.32614/CRAN.package.rStrava},
+  url     = {https://CRAN.R-project.org/package=rStrava}
+}
+
+@article{trackeR_jss,
+  author  = {Frick, Hannah and Kosmidis, Ioannis},
+  title   = {trackeR: Infrastructure for Running and Cycling Data from GPS-Enabled Tracking Devices in R},
+  journal = {Journal of Statistical Software},
+  year    = {2017},
+  volume  = {82},
+  number  = {7},
+  pages   = {1--29},
+  doi     = {10.18637/jss.v082.i07}
+}
+
+@Manual{activatr,
+  title   = {activatr: Utilities for Parsing and Plotting Activities},
+  author  = {Daniel Schafer},
+  year    = {2024},
+  note    = {R package version 0.2.1},
+  url     = {https://CRAN.R-project.org/package=activatr}
+}
+
+@Misc{FITfileR,
+  title        = {FITfileR: Read FIT Files in R},
+  author       = {Mike Smith},
+  year         = {2024},
+  howpublished = {GitHub repository},
+  url          = {https://github.com/grimbough/FITfileR},
+  note         = {Accessed 2025-10}
+}
+
+@Manual{injurytools,
+  title   = {injurytools: A Toolkit for Sports Injury Data Analysis},
+  author  = {Lore Zumeta Olaskoaga and Dae-Jin Lee},
+  year    = {2023},
+  note    = {R package version 1.0.3},
+  doi     = {10.32614/CRAN.package.injurytools},
+  url     = {https://CRAN.R-project.org/package=injurytools}
+}
