+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{gabbett2016,
+﻿@article{gabbett2016,
   author  = {Gabbett, Tim J.},
   title   = {The training–injury prevention paradox: should athletes be training smarter and harder?},
   journal = {British Journal of Sports Medicine},
