### 2025-10-11 04d3e21
Update paper.md
