### 2025-10-11 5c07763
Add Awesome badge to README
