### 2025-10-11 6ad6128
Add Awesome Sports Analytics badge to README
