### 2025-10-12 0d43319
Fix plot rendering and citation issues
- Fix Y-axis label overlap in plot_pbs() by correcting time formatting function
- Update DESCRIPTION author information to Zhiang He
- Add custom CITATION file for proper @Manual format
- Regenerate pkgdown site with corrected citation and plot fixes
