+++ DESCRIPTION
@@ -1,10 +1,10 @@
 Package: Athlytics
 Title: Academic R Package for Sports Physiology Analysis from Local 'Strava' Data
 Version: 1.0.0
-Author: Ang [aut, cre]
-Maintainer: Ang <ang@hezhiang.com>
+Author: Zhiang He [aut, cre]
+Maintainer: Zhiang He <ang@hezhiang.com>
 Authors@R: 
-    person("Ang", email = "ang@hezhiang.com", role = c("aut", "cre"))
+    person("Zhiang", "He", email = "ang@hezhiang.com", role = c("aut", "cre"))
 Description: An open-source computational framework for longitudinal analysis of 
     exercise physiology metrics using local 'Strava' data exports. Designed 
     for personal analysis and sports science applications, this package provides 
