### 2025-10-13 fe9e6d0
Add Professor Iztok Fister Jr. to acknowledgements
