### 2025-10-13 6c585c4
Update documentation: README description, paper acknowledgements, and plot fixes
