+++ R/flag_quality.R
@@ -37,7 +37,7 @@
 #'     sudden jumps (Delta HR > 10 bpm/s, Delta P > 300 W/s).
 #'   \item **GPS Drift**: Flags implausible speeds or accelerations based on sport type.
 #'   \item **Steady-State Detection**: Identifies segments with low variability
-#'     (CV < 8\%) lasting >= 20 minutes, suitable for EF/decoupling calculations.
+#'     (CV < 8%) lasting >= 20 minutes, suitable for EF/decoupling calculations.
 #' }
 #'
 #' The function is sport-aware and adjusts thresholds accordingly. All thresholds
