### 2025-10-13 cb4f757
Update README with new template and improved structure
