+++ README.md
@@ -1,81 +1,78 @@
 ﻿<div align="center">
 
-<img src="man/figures/image.png" alt="Athlytics Logo" width="200"/>
+<img src="man/figures/image.png" alt="Athlytics Logo" width="220"/>
 
 # Athlytics
-
 *An all-in-one, offline R toolkit for endurance analytics — from FIT/TCX/GPX & Strava exports to core models (ACWR, EF, Pa:Hr decoupling, PB, Exposure) with built-in QC & uncertainty.*
-
-[![Awesome Sports Analytics](https://awesome.re/badge-flat.svg)](https://github.com/firefly-cpp/awesome-computational-intelligence-in-sports?tab=readme-ov-file#software-)
-[![CRAN Status](https://img.shields.io/badge/CRAN-Athlytics-orange?style=flat-square)](https://cran.r-project.org/package=Athlytics)
-[![Total Downloads](https://cranlogs.r-pkg.org/badges/grand-total/Athlytics?color=blue&style=flat-square)](https://cran.r-project.org/package=Athlytics)
-[![R-CMD-check](https://img.shields.io/github/actions/workflow/status/HzaCode/Athlytics/R-CMD-check.yml?style=flat-square&label=R-CMD-check)](https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml)
-[![Code Coverage](https://img.shields.io/codecov/c/github/HzaCode/Athlytics?style=flat-square)](https://app.codecov.io/gh/HzaCode/Athlytics)
-[![MIT License](https://img.shields.io/badge/License-MIT-yellow.svg?style=flat-square)](https://opensource.org/licenses/MIT)
-[![Documentation](https://img.shields.io/badge/docs-passing-brightgreen?style=flat-square)](https://hezhiang.com/Athlytics/)
-
-**[📖 Documentation](https://hezhiang.com/Athlytics/)** • **[🚀 Quick Start](#quick-start)** • **[📊 Examples](#core-analyses)** • **[💬 Issues](https://github.com/HzaCode/Athlytics/issues)**
-
-<br>
-
-### 📈 Package Adoption & Growth
-
-<img src="man/figures/Athlytics_Final.png" alt="Athlytics Downloads Growth" width="650"/>
-
-*Used by researchers, coaches, and athletes*
 
 </div>
 
----
+<p align="center">
+    <a href="https://cran.r-project.org/web/views/SportsAnalytics.html"><img src="https://img.shields.io/badge/CRAN%20Listed-Sports%20Analytics-orange?style=flat-square" alt="CRAN Listed"></a>
+    <a href="https://cran.r-project.org/package/Athlytics"><img src="https://img.shields.io/badge/CRAN-Accepted-blue?style=flat-square" alt="CRAN Status"></a>
+    <a href="https://cran.r-project.org/package/Athlytics"><img src="https://cranlogs.r-pkg.org/badges/grand-total/Athlytics?color=blue&style=flat-square" alt="Total Downloads"></a>
+    <a href="https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml"><img src="https://img.shields.io/github/actions/workflow/status/HzaCode/Athlytics/R-CMD-check.yml?style=flat-square&label=R-CMD-check" alt="R-CMD-check"></a>
+    <a href="https://app.codecov.io/gh/HzaCode/Athlytics"><img src="https://img.shields.io/codecov/c/github/HzaCode/Athlytics?style=flat-square" alt="Code Coverage"></a>
[…]
