### 2025-10-13 796c5f3
Add Awesome Sports Analytics badge with correct link
