### 2025-10-14 3349c62
Improve README layout and navigation
- Consolidate navigation links in header, remove duplicate footer links
- Add Paper link to top navigation for higher visibility
- Fix anchor links by removing emoji prefixes
- Reorder badges: CRAN 鈫?Listed 鈫?Downloads 鈫?R-CMD-check 鈫?Coverage 鈫?License 鈫?Awesome
- Optimize images for mobile with better width and alt text
- Update installation section to reflect CRAN status with feature limitations
- Enhance content readability with bolded keywords
