+++ README.md
@@ -1,42 +1,59 @@
 ﻿<div align="center">
 
-<img src="man/figures/image.png" alt="Athlytics Logo" width="220"/>
+<img src="man/figures/image.png" alt="Athlytics logo" width="220"/>
 
 # Athlytics
 *An all-in-one, offline R toolkit for endurance analytics — from FIT/TCX/GPX & Strava exports to core models (ACWR, EF, Pa:Hr decoupling, PB, Exposure) with built-in QC & uncertainty.*
 
-</div>
-
-<p align="center">
-    <a href="https://github.com/firefly-cpp/awesome-computational-intelligence-in-sports?tab=readme-ov-file#software-"><img src="https://awesome.re/badge-flat.svg" alt="Awesome Sports Analytics"></a>
-    <a href="https://cran.r-project.org/web/views/SportsAnalytics.html"><img src="https://img.shields.io/badge/CRAN%20Listed-Sports%20Analytics-orange?style=flat-square" alt="CRAN Listed"></a>
-    <a href="https://cran.r-project.org/package/Athlytics"><img src="https://img.shields.io/badge/CRAN-Accepted-blue?style=flat-square" alt="CRAN Status"></a>
-    <a href="https://cran.r-project.org/package/Athlytics"><img src="https://cranlogs.r-pkg.org/badges/grand-total/Athlytics?color=blue&style=flat-square" alt="Total Downloads"></a>
-    <a href="https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml"><img src="https://img.shields.io/github/actions/workflow/status/HzaCode/Athlytics/R-CMD-check.yml?style=flat-square&label=R-CMD-check" alt="R-CMD-check"></a>
-    <a href="https://app.codecov.io/gh/HzaCode/Athlytics"><img src="https://img.shields.io/codecov/c/github/HzaCode/Athlytics?style=flat-square" alt="Code Coverage"></a>
-    <a href="https://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/License-MIT-yellow.svg?style=flat-square" alt="MIT License"></a>
+<p>
+  📘 <a href="https://hezhiang.com/Athlytics/"><strong>Docs</strong></a> &nbsp;&bull;&nbsp;
+  📦 <a href="https://cran.r-project.org/package=Athlytics"><strong>CRAN</strong></a> &nbsp;&bull;&nbsp;
+  🧾 <a href="https://github.com/HzaCode/Athlytics/blob/main/paper/paper.md"><strong>Paper</strong></a> &nbsp;&bull;&nbsp;
+  🐞 <a href="https://github.com/HzaCode/Athlytics/issues"><strong>Issues</strong></a>
 </p>
 
-<h3 align="center">
-  <a href="https://hezhiang.com/Athlytics/"><strong>📖 Docs</strong></a>
-  &nbsp;&nbsp;&bull;&nbsp;&nbsp;
-  <a href="#-quick-start"><strong>🚀 Quick Start</strong></a>
-  &nbsp;&nbsp;&bull;&nbsp;&nbsp;
-  <a href="#-core-analyses"><strong>📊 Core Analyses</strong></a>
-  &nbsp;&nbsp;&bull;&nbsp;&nbsp;
-  <a href="https://github.com/HzaCode/Athlytics/issues"><strong>💬 Issues</strong></a>
+<p>
+  <a href="https://cran.r-project.org/package=Athlytics">
+    <img src="https://img.shields.io/badge/CRAN-Accepted-blue?style=flat-square" alt="CRAN">
+  </a>
+  <a href="https://cran.r-project.org/web/views/SportsAnalytics.html">
[…]
