### 2025-10-14 4b37e4c
Clean up test files and improve test organization
