+++ tests/testthat/helper-mock-files.R
@@ -1,0 +1,211 @@
+# tests/testthat/helper-mock-files.R
+# Helper functions to create mock Strava export files
+
+library(xml2)
+
+# Create a temporary directory structure that mimics Strava export
+create_mock_strava_export <- function(base_dir = tempdir(), n_activities = 5) {
+  export_dir <- file.path(base_dir, "mock_strava_export")
+  activities_dir <- file.path(export_dir, "activities")
+  
+  # Create directories
+  dir.create(export_dir, showWarnings = FALSE, recursive = TRUE)
+  dir.create(activities_dir, showWarnings = FALSE)
+  
+  # Create activities.csv with proper column names (spaces, not dots)
+  activities_data <- data.frame(
+    `Activity ID` = seq_len(n_activities),
+    `Activity Date` = format(seq(Sys.Date() - n_activities + 1, Sys.Date(), by = "day"), "%b %d, %Y, %I:%M:%S %p"),
+    `Activity Name` = paste("Morning Run", seq_len(n_activities)),
+    `Activity Type` = "Run",
+    `Activity Description` = "",
+    `Elapsed Time` = round(runif(n_activities, 1800, 3600)),
+    Distance = round(runif(n_activities, 5000, 10000)),
+    `Max Heart Rate` = round(runif(n_activities, 160, 180)),
+    `Relative Effort` = round(runif(n_activities, 50, 100)),
+    Commute = "false",
+    `Activity Gear` = "",
+    Filename = paste0("activities/", seq_len(n_activities), ".tcx"),
+    `Athlete Weight` = "",
+    `Bike Weight` = "",
+    `Elapsed Time.1` = round(runif(n_activities, 1800, 3600)),
+    `Moving Time` = round(runif(n_activities, 1700, 3500)),
+    Distance.1 = round(runif(n_activities, 5000, 10000)),
+    `Max Speed` = round(runif(n_activities, 4, 6), 1),
+    `Average Speed` = round(runif(n_activities, 3, 4), 1),
+    `Elevation Gain` = round(runif(n_activities, 50, 200)),
+    `Elevation Loss` = round(runif(n_activities, 50, 200)),
+    `Elevation Low` = round(runif(n_activities, 50, 100)),
+    `Elevation High` = round(runif(n_activities, 100, 300)),
+    `Max Grade` = round(runif(n_activities, 5, 15)),
+    `Average Grade` = round(runif(n_activities, 0, 2), 1),
+    `Max Cadence` = round(runif(n_activities, 170, 190)),
+    `Average Cadence` = round(runif(n_activities, 160, 180)),
+    `Average Heart Rate` = round(runif(n_activities, 140, 160)),
+    Calories = round(runif(n_activities, 300, 600)),
+    `Max Heart Rate.1` = round(runif(n_activities, 160, 180)),
+    `Average Watts` = "",
+    `Max Watts` = "",
+    `Weighted Average Power` = "",
+    `Relative Effort.1` = round(runif(n_activities, 50, 100)),
+    `Average Elapsed Speed` = round(runif(n_activities, 2.8, 3.8), 1),
+    check.names = FALSE
+  )
+  
+  write.csv(activities_data, file.path(export_dir, "activities.csv"), row.names = FALSE)
+  
+  # Create mock activity files
+  for (i in seq_len(n_activities)) {
+    create_mock_tcx_file(
+      file.path(activities_dir, paste0(i, ".tcx")),
+      activity_date = Sys.Date() - n_activities + i,
+      duration_seconds = activities_data$`Elapsed Time`[i],
[…]
