+++ tests/testthat/test-calculate-functions.R
@@ -1,0 +1,340 @@
+# tests/testthat/test-calculate-functions.R
+# Tests for calculate functions with simulated data
+
+library(testthat)
+library(Athlytics)
+library(dplyr)
+library(lubridate)
+
+# Helper function to create realistic activity data
+create_realistic_activities <- function(n_days = 90, base_date = Sys.Date() - 90) {
+  dates <- seq(base_date, by = "day", length.out = n_days)
+  
+  # Simulate realistic running pattern (4-5 runs per week)
+  run_days <- sort(sample(1:n_days, size = round(n_days * 0.6)))
+  
+  activities <- data.frame(
+    id = seq_along(run_days),
+    name = paste("Morning Run", seq_along(run_days)),
+    type = "Run",
+    sport_type = "Run",
+    date = dates[run_days],
+    start_date_local = as.POSIXct(dates[run_days]),
+    distance = abs(rnorm(length(run_days), mean = 8000, sd = 2000)),  # 8km average
+    moving_time = abs(rnorm(length(run_days), mean = 2400, sd = 600)),  # 40min average
+    elapsed_time = abs(rnorm(length(run_days), mean = 2500, sd = 700)),
+    total_elevation_gain = abs(rnorm(length(run_days), mean = 100, sd = 50)),
+    average_heartrate = abs(rnorm(length(run_days), mean = 145, sd = 15)),
+    max_heartrate = abs(rnorm(length(run_days), mean = 175, sd = 10)),
+    average_speed = abs(rnorm(length(run_days), mean = 3.3, sd = 0.3)),
+    max_speed = abs(rnorm(length(run_days), mean = 4.5, sd = 0.5)),
+    average_cadence = rnorm(length(run_days), mean = 170, sd = 10),
+    average_watts = NA_real_,  # No power for runs
+    kilojoules = NA_real_,
+    has_heartrate = TRUE,
+    gear_id = NA_character_,
+    filename = paste0("activities/", run_days, ".fit"),
+    stringsAsFactors = FALSE
+  )
+  
+  # Ensure positive values
+  activities$distance <- abs(activities$distance)
+  activities$moving_time <- abs(activities$moving_time)
+  activities$elapsed_time <- pmax(activities$elapsed_time, activities$moving_time)
+  
+  # Add calculated fields
+  activities$distance_km <- activities$distance / 1000
+  activities$duration_mins <- activities$moving_time / 60
+  activities$average_speed_kmh <- (activities$distance_km / activities$duration_mins) * 60
+  
+  return(activities)
+}
+
+# Helper to create activity stream data
+create_activity_stream <- function(duration_seconds = 3600, steady_state = TRUE) {
+  time_points <- seq(0, duration_seconds, by = 1)
+  n_points <- length(time_points)
+  
+  # Base values
+  base_hr <- 150
+  base_pace <- 3.0  # m/s
+  base_power <- 200  # watts
+  
+  if (steady_state) {
+    # Steady state: low variation
+    hr_variation <- rnorm(n_points, 0, 2)
+    pace_variation <- rnorm(n_points, 0, 0.05)
+    power_variation <- rnorm(n_points, 0, 5)
+  } else {
+    # Non-steady state: high variation
+    hr_variation <- sin(time_points / 300) * 20 + rnorm(n_points, 0, 5)
+    pace_variation <- sin(time_points / 200) * 0.5 + rnorm(n_points, 0, 0.1)
+    power_variation <- sin(time_points / 250) * 50 + rnorm(n_points, 0, 10)
+  }
+  
+  stream_df <- data.frame(
+    time = time_points,
+    heartrate = round(base_hr + hr_variation),
+    velocity_smooth = base_pace + pace_variation,
+    distance = cumsum(base_pace + pace_variation),
[…]

[…]
