### 2025-10-14 948d3fa
Add advanced ACWR EWMA tests to improve coverage to 79.62%
