### 2025-10-15 6141658
docs: Update and refine README.md
Clarify project features and installation instructions.

Improve formatting for better readability.
