### 2025-10-15 923fb9a
Update paper acknowledgements and fix README citation anchor
