### 2025-10-16 30206e0
chore: clean up duplicate test files and update test configuration
