### 2025-10-16 25bd7b1
Fix R CMD check issues: remove duplicate data files, add FITfileR to Suggests, fix global variable '.' in load_local_activities, update .gitignore
