+++ DESCRIPTION
@@ -32,6 +32,7 @@ Imports:
     tools
 Suggests:
     devtools,
+    FITfileR,
     knitr,
     pkgdown,
     rmarkdown,
