### 2025-10-16 89baa1a
Fix CI: Remove FITfileR from Suggests as it's not on CRAN (GitHub-only package)
