### 2025-10-16 38f686b
Correct capitalization of 'Athlytics' in CONTRIBUTING.md
