### 2025-10-16 03e7bf9
Add comprehensive tests for calculate_ef_from_stream to boost coverage from 79.53% to 84.82%
