+++ tests/testthat/test-ef-stream-coverage.R
@@ -1,0 +1,264 @@
+# Test to boost coverage for calculate_ef_from_stream function
+
+test_that("calculate_ef_from_stream handles missing required columns", {
+  # Test missing velocity data for pace_hr
+  stream_no_velocity <- data.frame(
+    time = 1:100,
+    heartrate = rep(150, 100)
+  )
+  
+  result <- calculate_ef_from_stream(
+    stream_data = stream_no_velocity,
+    activity_date = Sys.Date(),
+    act_type = "Run",
+    ef_metric = "pace_hr"
+  )
+  
+  expect_equal(result$status, "missing_velocity_data")
+  expect_true(is.na(result$ef_value))
+  
+  # Test missing power data for power_hr
+  stream_no_power <- data.frame(
+    time = 1:100,
+    heartrate = rep(140, 100)
+  )
+  
+  result2 <- calculate_ef_from_stream(
+    stream_data = stream_no_power,
+    activity_date = Sys.Date(),
+    act_type = "Ride",
+    ef_metric = "power_hr"
+  )
+  
+  expect_equal(result2$status, "missing_power_data")
+  expect_true(is.na(result2$ef_value))
+  
+  # Test missing heartrate column
+  stream_no_hr <- data.frame(
+    time = 1:100,
+    distance = 1:100
+  )
+  
+  result3 <- calculate_ef_from_stream(
+    stream_data = stream_no_hr,
+    activity_date = Sys.Date(),
+    act_type = "Run",
+    ef_metric = "pace_hr"
+  )
+  
+  expect_equal(result3$status, "missing_hr_data")
+  expect_true(is.na(result3$ef_value))
+})
+
+test_that("calculate_ef_from_stream handles insufficient data", {
+  # Test insufficient data points (< 100)
+  small_stream <- data.frame(
+    time = 1:50,
+    heartrate = rep(150, 50),
+    distance = 1:50
+  )
+  
+  result <- calculate_ef_from_stream(
+    stream_data = small_stream,
+    activity_date = Sys.Date(),
+    act_type = "Run",
+    ef_metric = "pace_hr"
+  )
+  
+  expect_equal(result$status, "insufficient_data_points")
+  expect_true(is.na(result$ef_value))
+})
+
+test_that("calculate_ef_from_stream handles low HR coverage", {
+  # Create stream with lots of NA heartrate
+  stream_low_hr <- data.frame(
+    time = 1:200,
+    heartrate = c(rep(150, 30), rep(NA, 170)),
+    distance = 1:200 * 10
+  )
+  
+  result <- calculate_ef_from_stream(
+    stream_data = stream_low_hr,
+    activity_date = Sys.Date(),
+    act_type = "Run",
+    ef_metric = "pace_hr",
+    min_hr_coverage = 0.7  # Require 70% HR coverage
+  )
+  
+  expect_true(result$status %in% c("insufficient_hr_data", "insufficient_valid_data", "insufficient_data_points"))
+  expect_true(is.na(result$ef_value))
+})
+
+test_that("calculate_ef_from_stream handles velocity calculation from distance", {
+  # Test with distance column (no velocity_smooth)
+  stream_with_distance <- data.frame(
+    time = seq(0, 599, by = 1),  # 10 minutes of data
+    heartrate = rep(150, 600),
+    distance = seq(0, 3000, length.out = 600)  # 3km in 10 minutes
+  )
+  
+  result <- calculate_ef_from_stream(
+    stream_data = stream_with_distance,
+    activity_date = Sys.Date(),
+    act_type = "Run",
+    ef_metric = "pace_hr",
+    min_steady_minutes = 5,
+    steady_cv_threshold = 0.1,
+    min_hr_coverage = 0.7,
[…]
