### 2025-10-16 85d4138
Remove Chinese comments from test files, replace with English
