+++ tests/testthat/test-date-ranges.R
@@ -9,15 +9,15 @@ test_that("date range tests for coverage", {
     average_speed = c(10/3600*1000, 15/5400*1000, 20/7200*1000)
   )
   
-  # 测试开始日期
+  # Test start date filtering
   result1 <- calculate_ef(activities, start_date = "2023-03-01", end_date = "2024-01-01", quality_control = "off")
   expect_s3_class(result1, "data.frame")
   
-  # 测试结束日期
+  # Test end date filtering
   result2 <- calculate_ef(activities, start_date = "2022-01-01", end_date = "2023-09-01", quality_control = "off")
   expect_s3_class(result2, "data.frame")
   
-  # 测试日期范围
+  # Test date range filtering
   result3 <- calculate_ef(activities, start_date = "2023-02-01", end_date = "2023-08-01", quality_control = "off")
   expect_s3_class(result3, "data.frame")
 })
