+++ tests/testthat/test-parameter-boundaries.R
@@ -9,15 +9,15 @@ test_that("parameter boundary tests for coverage", {
     average_speed = 10/3600*1000
   )
   
-  # 测试最小持续时间
+  # Test minimum duration parameter
   result1 <- calculate_ef(activities, start_date = "2023-05-01", end_date = "2023-07-01", min_duration_mins = 0, quality_control = "off")
   expect_s3_class(result1, "data.frame")
   
-  # 测试最大持续时间
+  # Test maximum duration parameter
   result2 <- calculate_ef(activities, start_date = "2023-05-01", end_date = "2023-07-01", min_duration_mins = 1000, quality_control = "off")
   expect_s3_class(result2, "data.frame")
   
-  # 测试不同的质量控制设置
+  # Test different quality control settings
   result3 <- calculate_ef(activities, start_date = "2023-05-01", end_date = "2023-07-01", quality_control = "off")
   expect_s3_class(result3, "data.frame")
   
