### 2025-10-16 4dada20
Fix pkgdown: add calculate_ef_from_stream to reference index
