### 2025-10-16 d3af2fb
Fix FITfileR WARNING: use getFromNamespace to avoid undeclared import warnings
