### 2025-10-17 53df81f
Revise acknowledgements for clarity and conciseness
