### 2025-10-18 499e3a5
Remove paper link from README
Removed link to the paper from the README.
