### 2025-10-18 932d88c
Update README.md formatting and add newline
Fix formatting issues and add a newline at the end of the file.
