### 2025-10-18 ee6f2ec
Revise Quick Start section header formatting
Updated the Quick Start section header in README.md.
