### 2025-10-18 dcbad66
Update links in README for local navigation
