+++ README.md
@@ -37,9 +37,9 @@
 </p>
 
 <h3 align="center">
-  <a href="https://github.com/HzaCode/Athlytics/blob/main/README.md#quick-start"><strong>🚀 Quick Start</strong></a> &nbsp;&bull;&nbsp;
-  <a href="https://github.com/HzaCode/Athlytics/blob/main/README.md#core-analyses"><strong>📊 Core Analyses</strong></a> &nbsp;&bull;&nbsp;
-  <a href="https://github.com/HzaCode/Athlytics/blob/main/README.md#citation"><strong>📝 Citation</strong></a>
+  <a href="#-quick-start"><strong>🚀 Quick Start</strong></a> &nbsp;&bull;&nbsp;
+  <a href="#-core-analyses"><strong>📊 Core Analyses</strong></a> &nbsp;&bull;&nbsp;
+  <a href="#-citation"><strong>📝 Citation</strong></a>
 </h3>
 
 
