### 2025-10-18 e18ff60
Clean up README by removing separator
Removed unnecessary separator from README.
