### 2025-10-18 d7d4caa
Correct capitalization of 'Athlytics' in CONTRIBUTING.md
