### 2025-10-18 ef190c2
Correct capitalization of 'Athlytics' in CONTRIBUTING.md
