### 2025-10-18 805c924
Update wording for Strava data export instructions
