+++ README.md
@@ -99,7 +99,7 @@ remotes::install_github("HzaCode/Athlytics")
 
 1.  Navigate to **[Strava Settings → My Account](https://www.strava.com/settings/profile)**.
 2.  Under "Download or Delete Your Account," click **"Get Started"** and then **"Request Your Archive"**.
-3.  You will receive an email with a download link (this can take a few hours).
+3.  You will receive an email with a download link (it may take a little while).
 4.  Download the ZIP file (e.g., `export_12345678.zip`). **There is no need to unzip it.**
 
 ### 💻 Step 2: Load and Analyze (Cohort Example)
