### 2025-10-18 26572de
Export parse_activity_file and calculate_ef_from_stream functions, fix README YAML parsing, update documentation
