+++ NAMESPACE
@@ -13,11 +13,13 @@ export(calculate_acwr)
 export(calculate_acwr_ewma)
 export(calculate_decoupling)
 export(calculate_ef)
+export(calculate_ef_from_stream)
 export(calculate_exposure)
 export(calculate_pbs)
 export(cohort_reference)
 export(flag_quality)
 export(load_local_activities)
+export(parse_activity_file)
 export(plot_acwr)
 export(plot_acwr_comparison)
 export(plot_acwr_enhanced)
