### 2025-10-18 2bc35bf
Update README.md to academic version with acknowledgements
