+++ README.md
@@ -4,7 +4,7 @@
 <img src="man/figures/image.png" alt="Athlytics logo" width="220"/>
 
 # Athlytics
-*An all-in-one, offline R toolkit for endurance analytics — from FIT/TCX/GPX & Strava exports to core models (ACWR, EF, Pa:Hr decoupling, PB, Exposure) with built-in QC & uncertainty.*
+*An offline, reproducible R toolkit for endurance-exercise data: from FIT/TCX/GPX & Strava archives to validated models (ACWR, EF, pa:hr decoupling, PB, exposure), with built-in QC and uncertainty estimation.*
 
 <p>
   📘 <a href="https://hezhiang.com/Athlytics/"><strong>Docs</strong></a> &nbsp;&bull;&nbsp;
@@ -48,9 +48,9 @@
 
 ## 🎯 Overview
 
-**Athlytics** is a powerful R package for analyzing your endurance training data from **local Strava exports**. It provides a complete, **offline** toolkit to transform your raw data into meaningful insights about fitness, fatigue, and performance.
+**Athlytics** is a research-oriented R package for the longitudinal analysis of endurance training. It operates entirely on **local Strava exports** (or FIT/TCX/GPX files), avoiding API dependencies to ensure **privacy** and long-term **reproducibility**.
 
-Designed for athletes, coaches, and sports scientists, Athlytics standardizes the entire analysis process—from loading your data to generating publication-quality plots. Because it works **offline** without needing API keys, your data remains completely **private** and your workflows are **stable and reproducible** forever.
+The package standardizes the workflow from data ingestion and quality control to model estimation and uncertainty quantification. Implemented endpoints include **acute-to-chronic workload ratio (ACWR)**, **aerobic efficiency (EF)**, and **cardiovascular decoupling (pa:hr)**, alongside personal-best and exposure profiles suitable for **single-subject** and **cohort** designs. All functions return tidy data, facilitating statistical modeling and figure generation for academic reporting.
 
 <br>
 <div align="center">
@@ -62,17 +62,13 @@ Designed for athletes, coaches, and sports scientists, Athlytics standardizes th
 
 ## ✨ Key Features
 
-*   ✅ **100% Local and Private**: Works directly with your Strava ZIP export. No data is ever uploaded, and no internet connection or API keys are needed. You have full control over your information.
+*   ✅ **Reproducible by design** – Fully offline; no API keys. Deterministic pipelines suitable for longitudinal studies.
+*   ✅ **Validated metrics** – Implements ACWR, EF, and decoupling commonly used in exercise physiology; integrated **QC** checks.
+*   ✅ **Uncertainty-aware** – Functions return estimates with variance/intervals where applicable, enabling principled inference.
+*   ✅ **Cohort support** – Built-in helpers for multi-athlete datasets and percentile-band references.
+*   ✅ **Tidy outputs** – Consistent, analysis-ready tibbles for downstream modeling and figure pipelines.
 
-*   ✅ **All-in-One Analysis Pipeline**: Handles everything from loading and cleaning raw data to calculating advanced metrics and creating visualizations. No more patching together different tools.
-
-*   ✅ **Robust Scientific Metrics**: Implements widely used metrics for performance analysis, including Acute-to-Chronic Workload Ratio (ACWR), Aerobic Efficiency (EF), and Cardiovascular Decoupling, with built-in quality checks.
-
-*   ✅ **Powerful Cohort Analysis**: Purpose-built for comparing multiple athletes. Easily load data for a team or group and benchmark an individual's progress against the cohort.
-
[…]
