### 2025-10-18 d09e786
Update README: remove Reproducibility section and streamline Acknowledgements
