### 2025-10-18 4b4cb3b
Add docs-passing badge and remove emojis from README
