+++ README.md
@@ -1,4 +1,3 @@
-
 <div align="center">
 
 <img src="man/figures/image.png" alt="Athlytics logo" width="220"/>
@@ -7,9 +6,9 @@
 *An offline, reproducible R toolkit for endurance-exercise data: from FIT/TCX/GPX & Strava archives to validated models (ACWR, EF, pa:hr decoupling, PB, exposure), with built-in QC and uncertainty estimation.*
 
 <p>
-  📘 <a href="https://hezhiang.com/Athlytics/"><strong>Docs</strong></a> &nbsp;&bull;&nbsp;
-  📦 <a href="https://cran.r-project.org/package=Athlytics"><strong>CRAN</strong></a> &nbsp;&bull;&nbsp;
-  🐞 <a href="https://github.com/HzaCode/Athlytics/issues"><strong>Issues</strong></a>
+  <a href="https://hezhiang.com/Athlytics/"><strong>Docs</strong></a> &nbsp;|&nbsp;
+  <a href="https://cran.r-project.org/package=Athlytics"><strong>CRAN</strong></a> &nbsp;|&nbsp;
+  <a href="https://github.com/HzaCode/Athlytics/issues"><strong>Issues</strong></a>
 </p>
 
 <p>
@@ -25,6 +24,9 @@
   <a href="https://github.com/HzaCode/Athlytics/actions/workflows/R-CMD-check.yml">
     <img src="https://img.shields.io/github/actions/workflow/status/HzaCode/Athlytics/R-CMD-check.yml?style=flat-square&label=R-CMD-check" alt="R-CMD-check">
   </a>
+  <a href="https://hezhiang.com/Athlytics/">
+    <img src="https://img.shields.io/badge/docs-passing-brightgreen?style=flat-square" alt="Documentation">
+  </a>
   <a href="https://app.codecov.io/gh/HzaCode/Athlytics">
     <img src="https://img.shields.io/codecov/c/github/HzaCode/Athlytics?style=flat-square" alt="Coverage">
   </a>
@@ -37,16 +39,16 @@
 </p>
 
 <h3 align="center">
-  <a href="#-quick-start"><strong>🚀 Quick Start</strong></a> &nbsp;&bull;&nbsp;
-  <a href="#-core-analyses"><strong>📊 Core Analyses</strong></a> &nbsp;&bull;&nbsp;
-  <a href="#-citation"><strong>📝 Citation</strong></a>
+  <a href="#quick-start"><strong>Quick Start</strong></a> &nbsp;&bull;&nbsp;
+  <a href="#core-analyses"><strong>Core Analyses</strong></a> &nbsp;&bull;&nbsp;
+  <a href="#citation"><strong>Citation</strong></a>
 </h3>
 
 
 </div>
 
 
-## 🎯 Overview
+## Overview
 
 **Athlytics** is a research-oriented R package for the longitudinal analysis of endurance training. It operates entirely on **local Strava exports** (or FIT/TCX/GPX files), avoiding API dependencies to ensure **privacy** and long-term **reproducibility**.
 
@@ -60,9 +62,9 @@ The package standardizes the workflow from data ingestion and quality control to
 
 
 
-## ✨ Key Features
+## Key Features
 
-*   ✅ **Reproducible by design** – Fully offline; no API keys. Deterministic pipelines suitable for longitudinal studies.
+*   **Reproducible by design** – Fully offline; no API keys. Deterministic pipelines suitable for longitudinal studies.
 *   ✅ **Validated metrics** – Implements ACWR, EF, and decoupling commonly used in exercise physiology; integrated **QC** checks.
 *   ✅ **Uncertainty-aware** – Functions return estimates with variance/intervals where applicable, enabling principled inference.
[…]
