### 2025-10-18 6f73477
Update documentation and metadata for rOpenSci submission
