+++ paper/paper.bib
@@ -87,3 +87,11 @@
   doi     = {10.32614/CRAN.package.injurytools},
   url     = {https://CRAN.R-project.org/package=injurytools}
 }
+
+@Manual{ACWR,
+  title   = {ACWR: Acute Chronic Workload Ratio Calculation},
+  author  = {Ben Torvaney},
+  year    = {2024},
+  note    = {R package version 0.2.0},
+  url     = {https://CRAN.R-project.org/package=ACWR}
+}
