+++ .github/ISSUE_TEMPLATE/documentation.md
@@ -1,0 +1,35 @@
+---
+name: Documentation improvement
+about: Suggest improvements to documentation, examples, or vignettes
+title: '[DOCS] '
+labels: documentation
+assignees: ''
+
+---
+
+**Which documentation needs improvement?**
+- [ ] Function documentation (Rd files)
+- [ ] README
+- [ ] Vignettes
+- [ ] Website (pkgdown)
+- [ ] Code comments
+- [ ] Other: ___________
+
+**Location**
+Please specify where the documentation issue is:
+- Function name: 
+- File path: 
+- URL: 
+
+**Current documentation (if applicable)**
+Paste or describe the current documentation that needs improvement.
+
+**Suggested improvement**
+Describe what should be added, clarified, or corrected.
+
+**Why is this improvement needed?**
+Explain how this would help users understand or use Athlytics better.
+
+**Additional context**
+Add any other context, examples, or references here.
+
