+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,59 @@
+---
+name: Bug report
+about: Create a report to help us improve Athlytics
+title: '[BUG] '
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Load data with '...'
+2. Run function '....'
+3. See error
+
+**Minimal reproducible example**
+```r
+# Please provide a minimal reproducible example
+library(Athlytics)
+
+# Your code here
+```
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Actual behavior**
+What actually happened instead.
+
+**Error message**
+```
+Paste the complete error message here
+```
+
+**System information:**
+ - OS: [e.g., Windows 11, macOS 14.0, Ubuntu 22.04]
+ - R version: [e.g., 4.3.2]
+ - Athlytics version: [e.g., 1.0.0]
+ - Installation source: [CRAN or GitHub]
+
+**Session Info**
+<details>
+<summary>Click to expand</summary>
+
+```r
+# Please run sessionInfo() and paste the output here
+sessionInfo()
+```
+</details>
+
+**Additional context**
+Add any other context about the problem here (e.g., size of dataset, specific activity types, etc.)
+
+**Data Privacy Note**
+Please do NOT share any personal training data or identifiable information in this issue.
+
