+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,48 @@
+---
+name: Feature request
+about: Suggest an idea for Athlytics
+title: '[FEATURE] '
+labels: enhancement
+assignees: ''
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Use case**
+How would this feature be used? What problem does it solve?
+- For single-athlete analysis?
+- For cohort/team analysis?
+- For research purposes?
+
+**Expected function signature (if applicable)**
+```r
+# Example of how you envision using this feature
+result <- new_function(
+  data = activities,
+  parameter1 = value1,
+  parameter2 = value2
+)
+```
+
+**Additional context**
+Add any other context, screenshots, or references about the feature request here.
+
+**References**
+If this feature is based on published research or methodology, please cite the relevant papers:
+- Author et al. (Year). Title. Journal. DOI: ...
+
+**Implementation considerations**
+Are you willing to contribute to implementing this feature? (optional)
+- [ ] Yes, I can help with implementation
+- [ ] Yes, I can help with testing
+- [ ] Yes, I can help with documentation
+- [ ] I can provide example data for testing (de-identified)
+
