+++ README_test.md
@@ -1,0 +1,16 @@
+# Athlytics
+
+Test minimal README
+
+## Installation
+
+```r
+install.packages("Athlytics")
+```
+
+## Usage
+
+```r
+library(Athlytics)
+```
+
