### 2025-10-18 c021274
Remove unnecessary files and update gitignore
