### 2025-10-18 b9ee48b
Update README: soften Strava export time expectation
