### 2025-10-18 #702 @HzaCode
Hi @mpadge,

Thank you so much for your quick and very encouraging response.

I'm delighted to hear that the new framing for `Athlytics` aligns well with rOpenSci's scientific goals. I truly appreciate you taking the time to discuss this with the review team and for your willingness to find reviewers with expertise in sports analytics.

The information regarding the concurrent review process with JOSS is also incredibly helpful and provides a clear path forward. I will follow your suggestion and contact JOSS to pause the review there as soon as I hear about your team's decision to proceed.

I'll look forward to hearing from you or the Editor-in-Chief.

Thanks again for all your guidance.
