### 2025-10-19 0c34589
fix: improve package quality per rOpenSci standards
- Fixed LaTeX Unicode chars (伪鈫抋lpha, 鈦烩啋/, 鈮堚啋approx.)
- Fixed utils::getFromNamespace imports
- Added @return docs and @examples for all functions
- Updated LICENSE to DCF format
- Cleaned up hidden files and .Rbuildignore

鉁?Passed rOpenSci pkgcheck
鉁?R CMD check: 0 errors
鉁?84.8% test coverage
