+++ R/calculate_acwr_ewma.R
@@ -35,8 +35,8 @@
 #'     with configurable half-lives. More responsive to recent changes.
 #' }
 #'
-#' **EWMA Formula**: The smoothing parameter α is calculated from half-life: 
-#' `α = ln(2) / half_life`. The EWMA update is: `E_t = α × L_t + (1-α) × E_{t-1}`
+#' **EWMA Formula**: The smoothing parameter alpha is calculated from half-life: 
+#' `alpha = ln(2) / half_life`. The EWMA update is: `E_t = alpha * L_t + (1-alpha) * E_{t-1}`
 #' where L_t is daily load and E_t is the exponentially weighted average.
 #'
 #' **Confidence Bands**: When `ci = TRUE` and `method = "ewma"`, uses **moving-block
