+++ man/athlytics_palette_science.Rd
@@ -6,7 +6,16 @@
 \usage{
 athlytics_palette_science()
 }
+\value{
+A character vector of 8 hex color codes
+}
 \description{
 Classic palette inspired by Science magazine's figure guidelines.
 Conservative and widely accepted in scientific community.
+}
+\examples{
+# Get Science palette colors
+colors <- athlytics_palette_science()
+colors[1]  # Dark blue
+
 }

[…]
