### 2025-10-19 437af23
fix: handle GitHub dependency FITfileR in CI workflow
