+++ .github/workflows/R-CMD-check.yml
@@ -39,8 +39,19 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::rcmdcheck, any::covr, any::devtools, any::mockery, any::testthat
+          extra-packages: any::rcmdcheck, any::covr, any::devtools, any::mockery, any::testthat, any::remotes
           needs: check
+
+      - name: Install GitHub dependencies
+        run: |
+          if (!requireNamespace("remotes", quietly = TRUE)) install.packages("remotes")
+          # Install FITfileR from GitHub (in Suggests, so won't fail if install fails)
+          tryCatch({
+            remotes::install_github("grimbough/FITfileR", quiet = TRUE)
+          }, error = function(e) {
+            message("Note: FITfileR installation skipped (optional dependency)")
+          })
+        shell: Rscript {0}
 
       - name: Get package name and version
         id: pkg_info
