### 2025-10-19 4488e2b
RH
fix: move FITfileR to Remotes to resolve CI dependency issues
