### 2025-10-20 c5bc5da
update: add analysis images to README
