### 2025-10-20 57c9763
chore: remove temporary scripts from git tracking
