### 2025-10-20 3a0df4d
fix: restore run_coverage.R (needed for CI) and update .gitignore
