### 2025-10-20 a9fe678
chore: remove regenerate_figures.R from git tracking
