### 2025-10-21 65bb617
docs: add R-Universe installation method to README
