### 2025-10-21 4b0dcff
docs: add R-Universe badge to README
