### 2025-10-26 #702 @HzaCode
Hi @mpadge,

Thank you for the update, that's very kind of you. I'll gladly wait to hear from the team next week, so please don't feel any rush.

The conversation has been very helpful for me. It's prompted me to think more deeply about where `Athlytics` might fit within the broader R ecosystem.

My hope is that the package can offer a useful tool for researchers wanting to use Strava data in their scientific work. I felt that a key challenge is often contextualizing an individual's data, so the package focuses on **benchmarking** as a potential way to help with this.

By making it more straightforward to compare an individual against a reference cohort, I hope `Athlytics` can help bridge the gap from personal tracking to group-level scientific analysis. This was the main goal I had in mind when developing it.

Thank you again for your thoughtful guidance.
